% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_mydata_fup_red}
\alias{build_mydata_fup_red}
\title{Build Data Object for Fup RED Bayesian Model}
\usage{
build_mydata_fup_red(this.data, Physiological.Protein.Conc)
}
\arguments{
\item{this.data}{(Data Frame) Subset of data containing all observations of one test compound.}

\item{Physiological.Protein.Conc}{(Numeric) The assumed physiological protein concentration 
for plasma protein binding calculations.}
}
\value{
A named list to be passed into the Bayesian model.
}
\description{
Builds a list of arguments required for JAGS from subset of level-2 data frame. 
The list is used as an argument to JAGS during level-4 processing.
}
