% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_mydata_fup_uc}
\alias{build_mydata_fup_uc}
\title{Build Data Object for Fup UC Bayesian Model}
\usage{
build_mydata_fup_uc(MS.data, CC.data, T1.data, T5.data, AF.data)
}
\arguments{
\item{MS.data}{(Data Frame) Subset of data containing all observations of one test compound.}

\item{CC.data}{(Data Frame) Subset of data containing observations of calibration curves samples.}

\item{T1.data}{(Data Frame) Subset of data containing observations of Whole Plasma T1h Samples.}

\item{T5.data}{(Data Frame) Subset of data containing observations of Whole Plasma T5h Samples.}

\item{AF.data}{(Data Frame) Subset of data containing observations of Aqueous Fraction samples.}
}
\value{
A named list to be passed into the Bayesian model.
}
\description{
Builds a list of arguments required for JAGS from subset of level-2 data frame. 
The list is used as an argument to JAGS during level-4 processing.
}
