% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caco2_L0}
\alias{caco2_L0}
\title{Caco2 Example Data Set Level-0}
\format{
A level 0 data.frame with 48 rows 17 variables: \describe{
\item{\code{Compound}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.ID}}{Compound ID used in the laboratory}
\item{\code{Date}}{Date sample was acquired}
\item{\code{Sample}}{Sample Name}
\item{\code{Type}}{Type of Caco2 sample}
\item{\code{Compound.Conc}}{Test concentration (uM)}
\item{\code{Peak.Area}}{Peak area of analyte (target compound)}
\item{\code{ISTD.Peak.Area}}{Peak area of internal standard (pixels)}
\item{\code{ISTD.Name}}{Name of compound used as internal standard (ISTD)}
\item{\code{Analysis.Params}}{General description of chemical analysis method}
\item{\code{Level0.File}}{Name of data file from laboratory that was used to compile level0 data table)}
\item{\code{Level0.Sheet}}{Name of "sheet" (for Excel workbooks) from which the laboratory data were read)}
\item{\code{Direction}}{Direction of the Caco-2 permeability experiment}
\item{\code{Vol.Donor}}{The volume (in cm^3) of the donor portion of the Caco-2 experimental well}
\item{\code{Vol.Receiver}}{The volume (in cm^3) of the receiver portion of the Caco-2 experimental well}
\item{\code{Dilution.Factor}}{Number of times sample was diluted}
}
}
\usage{
caco2_L0
}
\description{
A subset of Caco2 data from a Caco2 data file provided by Cyprotex. This subset
contains data for 3 compounds from the Excel file. Direction, Vol.Donor, Vol.Receiver 
and Dilution.Factor are not pre-existing columns, but are manually filled with information found in the file.
}
\keyword{datasets}
