% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_red_base.R
\name{calc_fup_red_base}
\alias{calc_fup_red_base}
\title{Calculate fraction unbound in plasma from rapid equilibruim dialysis data}
\usage{
calc_fup_red_base(
  PPB.data,
  this.conc = 100,
  FILENAME = "BASE_Model_Results",
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  compound.col = "Compound.Name",
  area.col = "Area",
  series.col = "Series",
  type.col = "Sample.Type",
  compound.conc.col = "Nominal.Conc",
  cal.col = "Cal",
  dilution.col = "Dilution.Factor",
  istd.col = "ISTD.Area",
  istd.name.col = "ISTD.Name",
  istd.conc.col = "ISTD.Conc",
  nominal.test.conc.col = "Test.Target.Conc"
)
}
\arguments{
\item{PPB.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type.}

\item{this.conc}{The plasma protein concentration relative to physiologic
levels (default 100\%)}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "BASE_Model_Results")}

\item{TEMP.DIR}{An optional directory where file writing may be faster.}

\item{NUM.CHAINS}{The number of Markov Chains to use. This allows evaluation
of convergence according to Gelman and Rubin diagnostic.}

\item{NUM.CORES}{The number of processors to use (default 2)}

\item{sample.col}{Which column of PPB.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{lab.compound.col}{Which column of PPB.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{dtxsid.col}{Which column of PPB.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{date.col}{Which column of PPB.data indicates the laboratory measurment
date (Defaults to "Date")}

\item{compound.col}{Which column of PPB.data indicates the test compound
(Defaults to "Compound.Name")}

\item{area.col}{Which column of PPB.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate (Defaults to "Series")}

\item{type.col}{Which column of PPB.data indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{cal.col}{Which column of PPB.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{dilution.col}{Which column of PPB.data indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{istd.col}{Which column of PPB.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{istd.name.col}{Which column of PPB.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc.col}{Which column of PPB.data indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{nominal.test.conc.col}{Which column of PPB.data indicates the intended
test chemical concentration at time zero (Defaults to "Test.Target.Conc")}

\item{JAGS.PATH}{The file path to JAGS.}

\item{RANDOM.SEED}{The seed used by the random number generator 
(default 1111)}
}
\value{
A data.frame containing quunantiles of the Bayesian posteriors
}
\description{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Sample Blank (no chemical, just plasma) \tab Blank\cr
  Time zero chemical and plasma \tab T0\cr
  Equilibrium chemical in phosphate-buffered well (no plasma) \tab PBS\cr
  Equilibrium chemical in plasma well \tab Plasma\cr
}
}
\author{
John Wambaugh and Chantel Nicolas
}
