% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clint_cheminfo}
\alias{clint_cheminfo}
\title{Clint Chemical Information Example Data set}
\format{
A chemical info data.frame with 7 rows and 6 variables: \describe{
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard - CCD)}
\item{\code{Analyte Name}}{Name of the test analyte/compound and the name used by the laboratory}
\item{\code{Internal Standard}}{Name of the internal standard (ISTD)}
\item{\code{Mix}}{Mix used for the sample}
\item{\code{Compound}}{Name of the test analyte/compound}
\item{\code{Chem.Lab.ID}}{Compound as described in the chemistry laboratory}
}
}
\usage{
clint_cheminfo
}
\description{
The chemical ID mapping information from mass spectrometry measurements of intrinsic
hepatic clearance (Clint) for cryopreserved pooled human hepatocytes.
Chemicals were per- and poly-fluorinated alkyl substance
(PFAS) samples. The experiments were led by Dr.s Marci Smeltz and Barbara Wetmore 
\insertCite{smeltz2023plasma}{invitroTKstats}. This data set contains 7 unique compounds/chemicals.
}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}

\insertRef{smeltz2023plasma}{invitroTKstats}
}
\keyword{datasets}
