% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint_point.R
\name{decay}
\alias{decay}
\title{Decay Model}
\usage{
decay(time.hours, conc, cal, k_elim)
}
\arguments{
\item{time.hours}{(Numeric) Incubation Time (in hour).}

\item{conc}{(Numeric) Test compound concentration at time 0.}

\item{cal}{(Numeric) Calibration index.}

\item{k_elim}{(Numeric) Rate constant.}
}
\value{
Test compound concentration at time \code{t}.
}
\description{
This function calculates and returns test compound concentration at time \code{t}, using a model of
exponential decay with time \eqn{C(t) = C_0*e^{-mt}}, where \eqn{C_0} is the 
test compound concentration at time 0, \code{m} is a rate constant (argument \code{k_elim}),
and \code{t} is the incubation time in hour.
}
