% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_utils.R
\name{.check_all_miss_cols}
\alias{.check_all_miss_cols}
\title{Check if all the data is missing for specified columns.}
\usage{
.check_all_miss_cols(data, req.cols)
}
\arguments{
\item{data}{Data frame to check.}

\item{req.cols}{Column names that should be checked for whether all data is missing.}
}
\description{
This function checks for whether any of the specified columns
  are missing all of their data, either `NA` and/or `NULL`.
}
