% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_utils.R
\name{.check_char_cols}
\alias{.check_char_cols}
\title{Check the character columns are correctly of character class.}
\usage{
.check_char_cols(data, char.cols)
}
\arguments{
\item{data}{Data frame to check.}

\item{char.cols}{Column names that should be of the character class.}
}
\description{
Check the character columns are correctly of character class.
}
