% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_utils.R
\name{.check_no_miss_cols}
\alias{.check_no_miss_cols}
\title{Check there is no missing data for specified columns.}
\usage{
.check_no_miss_cols(data, req.cols, return.missing = FALSE)
}
\arguments{
\item{data}{Data frame to check.}

\item{req.cols}{Columns with required data.}

\item{return.missing}{Logical argument, if `TRUE` return rows missing data in
column (list or vector by column name). (Default is `FALSE`.)}
}
\description{
This function checks for whether any of the required columns
  have a data entry of `NA` or `NULL`.
}
