% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_utils.R
\name{.check_num_cols}
\alias{.check_num_cols}
\title{Check the numeric columns are correctly of numeric class.}
\usage{
.check_num_cols(data, num.cols)
}
\arguments{
\item{data}{Data frame to check.}

\item{num.cols}{Column names that should be of the numeric class.}
}
\description{
Check the numeric columns are correctly of numeric class.
}
