% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_utils.R
\name{.check_std_colnames_in_data}
\alias{.check_std_colnames_in_data}
\title{Check the standard column names are in the data.}
\usage{
.check_std_colnames_in_data(data, std.colnames, data.name = NULL)
}
\arguments{
\item{data}{Data frame to check.}

\item{std.colnames}{Vector of character strings with standard column names
to check for in the data.}

\item{data.name}{Name of the data object passed to the standard column names
check function. (Defaults to NULL.)}
}
\description{
Check the standard column names are in the data.
}
