% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fup_red_L0}
\alias{fup_red_L0}
\title{Fup RED Level-0 Example Data set}
\format{
A level-0 data.frame with 660 rows and 18 variables: \describe{
\item{\code{Compound}}{Name of the test analyte/compound}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard - CCD)}
\item{\code{Lab.Compound.ID}}{Compound as described in the laboratory}
\item{\code{Date}}{Date the sample was added to the MS analyzer}
\item{\code{Sample}}{Sample description used in the laboratory}
\item{\code{Type}}{Type of RED sample, annotated by the laboratory}
\item{\code{Compound.Conc}}{Expected (or nominal) concentration of analyte (for calibration curve)}
\item{\code{Peak.Area}}{Peak area of analyte (target compound)}
\item{\code{ISTD.Peak.Area}}{Peak area of internal standard (ISTD) compound (pixels)}
\item{\code{ISTD.Name}}{Name of the internal standard (ISTD) analyte/compound}
\item{\code{Analysis.Params}}{Column contains the retention time}
\item{\code{Level0.File}}{Name of the laboratory data file from which the level-0 sample data was extracted}
\item{\code{Level0.Sheet}}{Name of the Excel workbook 'sheet' from which the level-0 sample data was extracted}
\item{\code{Sample Text}}{Additional notes on the sample}
\item{\code{Sample.Type}}{Type of RED sample in \code{invitroTKstats} package annotations}
\item{\code{Replicate}}{Identifier for repeated measurements of one sample of a compound}
\item{\code{Time}}{Time when the sample was measured - in hours (h)}
\item{\code{Dilution.Factor}}{Number of times the sample was diluted}
}
}
\usage{
fup_red_L0
}
\description{
Mass Spectrometry measurements of plasma protein binding (PPB) via rapid 
equilibrium dialysis (RED) for per- and poly-fluorinated alkyl substance
(PFAS) samples. Experiments were led by Dr.s Marci Smeltz and Barbara Wetmore 
\insertCite{smeltz2023plasma}{invitroTKstats}.
This data set is a subset of experimental data containing samples for 
3 test analytes/compounds.
}
\references{
\insertRef{waters2008validation}{invitroTKstats}

\insertRef{smeltz2023plasma}{invitroTKstats}
}
\keyword{datasets}
