% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fup_red_L4}
\alias{fup_red_L4}
\title{Fup RED Level-4 Example Data set}
\format{
A level-4 data.frame with 3 rows and 7 variables: \describe{
\item{\code{Compound.Name}}{Name of the test analyte/compound}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard - CCD)}
\item{\code{Fup.point}}{Point estimate of fraction unbound in plasma}
\item{\code{Fup.Med}}{Median fraction unbound in plasma}
\item{\code{Fup.Low}}{2.5th quantile of fraction unbound in plasma}
\item{\code{Fup.High}}{97.5th quantile of fraction unbound in plasma}
}
}
\usage{
fup_red_L4
}
\description{
Mass Spectrometry measurements of plasma protein binding (PPB) via rapid 
equilibrium dialysis (RED) for per- and poly-fluorinated alkyl substance
(PFAS) samples. Experiments were led by Dr.s Marci Smeltz and Barbara Wetmore 
\insertCite{smeltz2023plasma}{invitroTKstats}.
This data set is a subset of experimental data containing samples for 
3 test analytes/compounds.
}
\references{
\insertRef{waters2008validation}{invitroTKstats}

\insertRef{smeltz2023plasma}{invitroTKstats}
}
\keyword{datasets}
