% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fup_uc_cheminfo}
\alias{fup_uc_cheminfo}
\title{Fup UC Chemical Information Example Data set}
\format{
A chemical info data.frame with 75 rows and 4 variables: \describe{
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard - CCD)}
\item{\code{Chemical Name (Common Abbreviation)}}{Name of the test analyte/compound and abbreviation used by the lab as the compound ID}
\item{\code{Compound}}{Name of the test analyte/compound}
\item{\code{Chem.Lab.ID}}{Common abbreviation of the test analyte/compound as described in the laboratory}
}
}
\usage{
fup_uc_cheminfo
}
\description{
The chemical ID mapping information from mass spectrometry measurements of plasma protein binding (PPB) via 
ultracentrifugation (UC) for per- and poly-fluorinated alkyl substance
(PFAS) samples. Experiments were led by Dr.s Marci Smeltz and Barbara Wetmore 
\insertCite{smeltz2023plasma}{invitroTKstats}. 
This data set contains 75 unique compounds/chemicals.
}
\references{
\insertRef{howard2010plasma}{invitroTKstats}

\insertRef{smeltz2023plasma}{invitroTKstats}
}
\keyword{data}
