% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fup_uc_L3}
\alias{fup_uc_L3}
\title{Fup UC Level-3 Example Data set}
\format{
A level-3 data.frame with 3 rows and 5 variables: \describe{
\item{\code{Compound.Name}}{Name of the test analyte/compound}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard - CCD)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Calibration}}{Identifier for mass spectrometry calibration -- usually the date}
\item{\code{Fup}}{Fraction unbound in plasma}
}
}
\usage{
fup_uc_L3
}
\description{
Mass Spectrometry measurements of plasma protein binding (PPB) via 
ultracentrifugation (UC) for per- and poly-fluorinated alkyl substance
(PFAS) samples. Experiments were led by Dr.s Marci Smeltz and Barbara Wetmore 
\insertCite{smeltz2023plasma}{invitroTKstats}. 
This data set is a subset of experimental data containing samples for 
3 test analytes/compounds.
}
\references{
\insertRef{howard2010plasma}{invitroTKstats}

\insertRef{smeltz2023plasma}{invitroTKstats}
}
\keyword{data}
