% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kreutz2020}
\alias{kreutz2020}
\alias{Kreutz2020}
\title{Kreutz 2020 Ultracentrifucation Data Set}
\format{
A data.frame with 228 rows and 19 variables: \describe{
\item{\code{Lab.Sample.Name}}{}
\item{\code{Date}}{}
\item{\code{Compound.Name}}{}
\item{\code{DTXSID}}{}
\item{\code{Lab.Compound.Name}}{}
\item{\code{Sample.Type}}{}
\item{\code{Dilution.Factor}}{}
\item{\code{Cal}}{}
\item{\code{Standard.Conc}}{}
\item{\code{Test.Target.Conc}}{}
\item{\code{ISTD.Name}}{}
\item{\code{ISTD.Conc}}{}
\item{\code{IS.Area}}{}
\item{\code{Series}}{}
\item{\code{Area}}{}
\item{\code{Response}}{}
\item{\code{Analysis.Method}}{}
\item{\code{Analysis.Instrument}}{}
\item{\code{Analysis.Parameters}}{}
  }
}
\usage{
kreutz2020
}
\description{
Mass Spectrometry measurments of plasma protein binding measured by 
ultracentrifucation for per- and poly-fluorinated alkyl subtance
(PFAS) samples from experiments conducted by Dr. Anna Kreutz
}
\references{
\insertRef{howard2010plasma}{invitroTKstats}
}
\keyword{data}
