% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L1.common.col}
\alias{L1.common.col}
\alias{L1.common.cols}
\title{Common Columns in Level-1}
\format{
A named character vector containing the default/standard column names
across HTTK assays, where the element names are the corresponding L1 arguments.
}
\usage{
L1.common.cols
}
\description{
Common column names across the various \emph{in vitro} assays used for collecting
\emph{in vitro} toxicokinetic parameters.
}
\keyword{datasets}
