% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint_point.R
\name{lldecay}
\alias{lldecay}
\title{Log-Likelihood of residual function}
\usage{
lldecay(this.data, cal, k_elim, sigma)
}
\arguments{
\item{this.data}{(Data Frame) A data frame containing data of a compound.}

\item{cal}{(Numeric) Calibration index.}

\item{k_elim}{(Numeric) Rate constant.}

\item{sigma}{(Numeric) standard deviation of the residual}
}
\value{
Negative log-likelihood.
}
\description{
This function returns the negative log-likelihood
}
