% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint_point.R
\name{llsatdecay}
\alias{llsatdecay}
\title{Log-Likelihood Function}
\usage{
llsatdecay(this.data, cal, k_elim, sigma, sat)
}
\arguments{
\item{this.data}{(Data Frame) A data frame containing data of a compound.}

\item{cal}{(Numeric) Calibration}

\item{k_elim}{(Numeric) Rate constant.}

\item{sigma}{(Numeric) standard deviation.}

\item{sat}{(Numeric) saturation probability.}
}
\description{
This function returns the negative log-likelihood.
}
