% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uc_results.R
\name{plot_uc_results}
\alias{plot_uc_results}
\title{Plot Measurements of Fraction Unbound in Plasma (UC) of a Compound}
\usage{
plot_uc_results(
  dat,
  bayes,
  compound,
  cal,
  MW,
  quad.cal = NULL,
  cal.col = "Cal",
  name.col = "Compound.Name",
  uc.dilute = 5,
  af.dilute = 2
)
}
\arguments{
\item{dat}{level 2 data set}

\item{bayes}{MCMC results}

\item{compound}{name of a compound}

\item{cal}{Calibration Identifier}

\item{MW}{molecular weight}

\item{quad.cal}{Quadratic Calibration curve results from Mass Spec
(Defaults to NULL)}

\item{cal.col}{Which column of data indicates the calibration
(Defaults to "Cal")}

\item{name.col}{Which column of data indicates the test compound
(Defaults to "Compound.Name")}

\item{uc.dilute}{ultracentrifugation dilution factor}

\item{af.dilute}{aqueous fraction dilution factor}
}
\value{
A list of two plots, one with axes in original unit and one's axes in log10-scale
}
\description{
This function plots the measurements of a compound with calibration curves
from ultracentrifugation data. Users can use the returned plots for a reality
check on the data collected and the Bayesian results, see if they are as expected.
Data are read from a "Level2" text file.
}
\author{
John Wambaugh
}
