% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint_point.R
\name{satdecay}
\alias{satdecay}
\title{Decay Formula with saturation}
\usage{
satdecay(time.hours, conc, cal, k_elim, sat)
}
\arguments{
\item{time.hours}{(Numeric) Incubation time in hours.}

\item{conc}{(Numeric) Concentration at time 0.}

\item{cal}{(Numeric) Calibration.}

\item{k_elim}{(Numeric) rate constant.}

\item{sat}{(Numeric) saturation probability, equals 1 when concentration is 10 uM.}
}
\description{
This function calculates and returns test compound concentration at time \code{t}, using a model of
exponential decay with time, in addition, taking into account of saturation
probability \eqn{C(t) = C_0*e^{-m*sat*t}}, where \eqn{C_0} is the test compound concentration at time 0, \code{m} is a rate constant (argument \code{k_elim}),
sat is the saturation probability, and \code{t} is the incubation time in hour.
saturation meaning observing a lower clearance at a higher concentration. Metabolism 
was completely saturated by 10 uM. (John's paper cite here)
}
