% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scientific_10}
\alias{scientific_10}
\title{Formatting function for X-axis in log10-scale}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{(Character) String to be formatted.}
}
\value{
Text with desired expression. Replace any scientific e notation to ten notation, 
simplify 10^01 to 10 and 10^0 to 1.
}
\description{
Formatting function for X-axis in log10-scale
}
