% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std.catcols}
\alias{std.catcols}
\title{Standard Data Catalog (Data Guide) Columns}
\format{
A named character vector containing the default/standard column names
for data catalogs, where the element names are the corresponding `create_catalog`
arguments.
}
\usage{
std.catcols
}
\description{
Standardized column names for data catalogs (i.e. data guides) used for
collecting the minimum information to merge level-0 data files.
}
\keyword{datasets}
