% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TO1caco2}
\alias{TO1caco2}
\alias{to1caco2}
\title{Cyprotex 2021 Caco2 Data Set}
\format{
A data.table with 3110 rows 18 variables: \describe{
\item{\code{SampleName}}{}
\item{\code{CompoundName}}{}
\item{\code{Feature}}{}          
\item{\code{Area}}{}
\item{\code{ISTD.Area}}{}
\item{\code{ISTDResponseRatio}}{}
\item{\code{TO}}{}
\item{\code{FileName}}{}
\item{\code{SheetName}}{}        
\item{\code{Type}}{}
\item{\code{Dilution.Factor}}{}
\item{\code{Direction}}{}       
\item{\code{Vol.Receiver}}{}
\item{\code{Vol.Donor}}{}
\item{\code{Date}}{}            
\item{\code{ISTD.Name}}{}
\item{\code{ISTD.Conc}}{}
\item{\code{Test.Target.Conc}}{} 
  }
}
\usage{
TO1caco2
}
\description{
Mass Spectrometry measurments of Caco2 membrane permeability as measured by
Cyprotex, a contractor to the U.S. EPA.
}
\references{
\insertRef{hubatsch2007determination}{invitroTKstats}
}
\keyword{data}
