% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.methods}
\alias{wambaugh2019.methods}
\alias{Wambaugh2019.methods}
\title{Mass Spectrometry Methods for \insertCite{wambaugh2019assessing;textual}{invitroTKstats} Chemicals}
\format{
A data.frame 520 rows and 14 variables: \describe{
\item{\code{DTXSID}}{}
\item{\code{PREFERRED_NAME}}{}
\item{\code{CASRN}}{}
\item{\code{MOLECULAR_FORMULA}}{}
\item{\code{AVERAGE_MASS}}{}
\item{\code{QSAR_READY_SMILES}}{}
\item{\code{LC}}{}
\item{\code{Agilent.QQQ}}{}
\item{\code{Water.s.Xevo}}{}
\item{\code{AB.Sciex.Qtrap}}{}
\item{\code{GC}}{}
\item{\code{Agilent.GCMS}}{}
\item{\code{GCTOF}}{}
\item{\code{Comment}}{}
  }
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.methods
}
\description{
As reported in \insertCite{wambaugh2019assessing;textual}{invitroTKstats}, contractor Cyprotex performed in vitro
toxicokinetic experiments requiring the development of analytical chemistry
methods to quantitative concentration. 
Liquid chromatography/mass spectrometry (LC/MS) was faster and 
attempted first. Gas chromatography/mass spectrometry (GC/MS) was attempted
when LC/MS failed. For some chemicals the attempts to develop an
analysis method failed for both LC/MS and GC/MS and no
in vitro experiments were conducted.
}
\references{
\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\keyword{data}
