% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.red}
\alias{wambaugh2019.red}
\alias{Wambaugh2019.red}
\title{Mass Spectrometry for \insertCite{wambaugh2019assessing;textual}{invitroTKstats} Protein Binding}
\format{
A data.frame 17,689 rows and 14 variables: \describe{
\item{\code{Preferred.Name}}{Preferred name used for a test compound.}
\item{\code{CAS}}{CAS Registry Number of the test compound.}
\item{\code{DTXSID}}{EPA's DSSTox Structure ID.}
\item{\code{SampleName}}{Sample name used by the laboratory.}
\item{\code{CompoundName}}{}
\item{\code{Transition}}{}
\item{\code{Area}}{The test compound MS peak area.}
\item{\code{ISTD.Area}}{MS peak area for the internal standard.}
\item{\code{ISTDResponseRatio}}{}
\item{\code{Dilution.Factor}}{How many times a sample was diluted.}
\item{\code{Protein}}{}
\item{\code{Task.Order}}{}
\item{\code{RawDataSet}}{Name of the file the data came from.}
\item{\code{CyprotexEPASetNumber}}{}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.red
}
\description{
\insertCite{wambaugh2019assessing;textual}{invitroTKstats} includes chemical-specific measurements of plasma
protein binding using the method of rapid equilibriuim dialysis
\insertCite{waters2008validation}{invitroTKstats}.
}
\references{
\insertRef{waters2008validation}{invitroTKstats}

\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\keyword{data}
