% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_mydata_fup_uc}
\alias{build_mydata_fup_uc}
\title{Build Data Object for Fup UC Bayesian Model}
\usage{
build_mydata_fup_uc(MS.data)
}
\arguments{
\item{MS.data}{(Data Frame) Subset of data containing all observations of one test compound.}
}
\value{
A named list to be passed into the Bayesian model.
}
\description{
Build Data Object for Fup UC Bayesian Model
}
