% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caco2_L1}
\alias{caco2_L1}
\title{Caco2 Example Data Set Level-1}
\format{
A level 1 data.frame with 48 rows 24 variables: \describe{
\item{\code{Lab.Sample.Name}}{Sample name as described in the laboratory}
\item{\code{Date}}{Date sample was acquired}
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Sample.Type}}{Type of Caco2 sample}
\item{\code{Direction}}{Direction of the Caco-2 permeability experiment}
\item{\code{Dilution.Factor}}{Number of times sample was diluted}
\item{\code{Calibration}}{Identifier for mass spectrometry calibration -- usually the date}
\item{\code{Series}}{Identier for replicate series of measurements}
\item{\code{Standard.Conc}}{Concentration of analytic standard (for calibration curve) (uM)}
\item{\code{Test.Target.Conc}}{Test concentration of chemical (uM)}
\item{\code{Time}}{Time point sample measured (h)}
\item{\code{ISTD.Name}}{Name of compound used as internal standard (ISTD)}
\item{\code{ISTD.Conc}}{Concentration of ISTD (uM)}
\item{\code{ISTD.Area}}{Peak area of internal standard (pixels)}
\item{\code{Area}}{Peak area of analyte (target compound)}
\item{\code{Membrane.Area}}{The area of the Caco-2 monolayer.}
\item{\code{Vol.Donor}}{The volume (in cm^3) of the donor portion of the Caco-2 experimental well}
\item{\code{Vol.Receiver}}{The volume (in cm^3) of the receiver portion of the Caco-2 experimental well}
\item{\code{Analysis.Method}}{General description of chemical analysis method}
\item{\code{Analysis.Instrument}}{Instrument(s) used for chemical analysis)}
\item{\code{Analysis.Parameters}}{Parameters for identifing analyte peak (for example, retention time)}
\item{\code{Response}}{Response factor (calculated from analyte and ISTD peaks)}
}
}
\usage{
caco2_L1
}
\description{
A subset of Caco2 data from a Caco2 data file provided by Cyprotex. This subset
contains data for 3 compounds from the Excel file and had been processed into a standard 
format with \code{format_caco2}.
}
\keyword{datasets}
