% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_caco2_point.R
\name{calc_caco2_point}
\alias{calc_caco2_point}
\title{Calculate a point estimate of apparent membrane permeability from Caco-2 data}
\usage{
calc_caco2_point(FILENAME, good.col = "Verified")
}
\arguments{
\item{FILENAME}{A string used to identify the input file, whatever the
argument given, "-Caco-2-Level2.tsv" is appended (defaults to "MYDATA")}

\item{good.col}{Name of a column indicating which rows have been verified for 
analysis, indicated by a "Y" (Defaults to "Verified")}
}
\value{
\item{data.frame}{A data.frame in standardized format} 
\tabular{rrr}{
  C0_A2B \tab Time zero donor concentration \tab Mass Spec Response Ratio (RR) \cr
  dQdt_A2B \tab Estimated rate of mass movement through membrane \tab RR*cm^3/s \cr
  Papp_A2B \tab Apparent membrane permeability \tab 10^-6 cm/s\cr
  C0_B2A \tab Time zero donor concentration \tab Mass Spec Response Ratio (RR) \cr
  dQdt_B2A \tab Estimated rate of mass movement through membrane \tab RR*cm^3/s \cr
  Papp_B2A \tab Apparent membrane permeability \tab 10^-6 cm/s\cr
  Refflux \tab Efflux ratio \tab unitless\cr
}
}
\description{
This function uses mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of membrane 
permeability using Caco-2 cells (Hubatsch, 2007).
Data are read from a "Level2" text file that should have been formatted and 
created 
by \code{\link{format_caco2}} (this is the "Level1" file). The Level1 file
should have been curated by adding a column with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).
}
\details{
The data frame of observations should be annotated according to direction 
(either apical to basal -- "AtoB" -- or basal to apical -- "BtoA") and type
of concentrtion measured:
\tabular{rr}{
  Blank with no chemical added \tab Blank \cr
  Dosing vehicle (C0) at target concentration \tab D0\cr
  Donor compartment at end of experiment \tab D2\cr
  Receiver compartment at end of experiment\tab R2\cr
}

Apparent membrane permeability (Papp) is calculated from MS responses as:

P_app = (dQ/dt) / C_0 / A

The rate of permeation, dQ/dt (peak area/s) is calculated as:

dQ/dt = max(0,(mean(PBS Response * Dilution.Factor) - 
  mean(Blank Response * Dilution.Factor)))
}
\examples{
library(invitroTKstats)
level0 <- TO1caco2
level1 <- format_caco2(level0,
  FILENAME="EPACyprotex2021",
  compound.col="CompoundName",
  compound.conc.col="Standard.Conc",
  membrane.area.col=0.11
  )

level2 <- level1
level2$Verified <- "Y"

write.table(level2,
  file="EPACyprotex2021-Caco-2-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)
  
level3 <- calc_caco2_point(FILENAME="EPACyprotex2021")

}
\references{
Hubatsch, Ina, Eva GE Ragnarsson, and Per Artursson. 
"Determination of drug permeability and prediction of drug absorption in 
Caco-2 monolayers." Nature protocols 2.9 (2007): 2111.
}
\author{
John Wambaugh
}
