% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_caco2_point.R
\name{calc_caco2_point}
\alias{calc_caco2_point}
\title{Calculate a point estimate of apparent membrane permeability from Caco-2 data}
\usage{
calc_caco2_point(
  FILENAME,
  data.in,
  good.col = "Verified",
  output.res = TRUE,
  sig.figs = 3,
  INPUT.DIR = NULL,
  OUTPUT.DIR = NULL
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the input Level-2 file.
"<FILENAME>-Caco-2-Level2.tsv".}

\item{data.in}{(Data Frame) A Level-2 data frame generated from the 
\code{format_caco2} function with a verification column added by 
\code{sample_verification}. Complement with manual verification if needed.}

\item{good.col}{(Character) Column name indicating which rows have been
verified, data rows valid for analysis are indicated with a "Y".
(Defaults to "Verified".)}

\item{output.res}{(Logical) When set to \code{TRUE}, the result 
table (Level-3) will be exported the current directory as a .tsv file. 
(Defaults to \code{TRUE}.)}

\item{sig.figs}{(Numeric) The number of significant figures to round the exported result table (Level-3). 
(Note: console print statements are also rounded to specified significant figures.) 
(Defaults to \code{3}.)}

\item{INPUT.DIR}{(Character) Path to the directory where the input level-2 file exists. 
If \code{NULL}, looking for the input level-2 file in the current working
directory. (Defaults to \code{NULL}.)}

\item{OUTPUT.DIR}{(Character) Path to the directory to save the output file. 
If \code{NULL}, the output file will be saved to the current working
directory or \code{INPUT.DIR} if specified. (Defaults to \code{NULL}.)}
}
\value{
\item{data.frame}{A data.frame in standardized format}
\tabular{rrr}{
  C0_A2B \tab Time zero donor concentration \tab Mass Spec Response Ratio (RR) \cr
  dQdt_A2B \tab Estimated rate of mass movement through membrane \tab RR*cm^3/s \cr
  Papp_A2B \tab Apparent membrane permeability \tab 10^-6 cm/s\cr
  C0_B2A \tab Time zero donor concentration \tab Mass Spec Response Ratio (RR) \cr
  dQdt_B2A \tab Estimated rate of mass movement through membrane \tab RR*cm^3/s \cr
  Papp_B2A \tab Apparent membrane permeability \tab 10^-6 cm/s\cr
  Refflux \tab Efflux ratio \tab unitless\cr
  Frec_A2B.vec \tab Fraction recovered for the apical-basal direction, calculated as the fraction of the initial donor amount recovered in the receiver compartment \tab unitless \cr
  Frec_B2B.vec \tab Fraction recovered for the basal-apical direction, calculated in the same way as Frec_A2B.vec but in the opposite transport direction \tab unitless \cr 
  Recovery_Class_A2B \tab Recovery classification for apical-to-basal permeability("Low Recovery" if Frec_A2B.vec < 0.4 or "High Recovery" if Frec_A2B.vec > 2.0) \tab qualitative category \cr
  Recovery_Class_B2A \tab Recovery classification for basal-to-apical permeability("Low Recovery" if Frec_B2A.vec < 0.4 or "High Recovery" if Frec_B2A.vec > 2.0) \tab qualitative category \cr
}
}
\description{
This function uses mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of membrane
permeability using Caco-2 cells \insertCite{hubatsch2007determination}{invitroTKstats}.
Data are read from a "Level2" text file that should have been formatted and
created
by \code{\link{format_caco2}} (this is the "Level1" file). The Level1 file
should have been curated by adding a column with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).
}
\details{
The data frame of observations should be annotated according to direction
(either apical to basal -- "AtoB" -- or basal to apical -- "BtoA") and type
of concentration measured:
\tabular{rr}{
  Blank with no chemical added \tab Blank \cr
  Dosing vehicle (C0) at target concentration \tab D0\cr
  Donor compartment at end of experiment \tab D2\cr
  Receiver compartment at end of experiment\tab R2\cr
}

Apparent membrane permeability (\eqn{P_{app}}) is calculated from MS responses as:


\eqn{P_{app} = \frac{dQ/dt}{c_0*A}}

The rate of permeation, \eqn{\frac{dQ}{dt}}\eqn{\left(\frac{\text{peak area}}{\text{time (s)}} \right)} is calculated as:

\eqn{\frac{dQ}{dt} = \max\left(0, \frac{\sum_{i=1}^{n_P} (r_P * c_{DF})}{n_P} - \frac{\sum_{i=1}^{n_B} (r_B * c_{DF})}{n_B}\right)}

where \eqn{r_P} is PBS Response, \eqn{c_{DF}} is Dilution Factor, \eqn{r_B} is Blank Response,
\eqn{n_P} is the number of PBS Responses, and \eqn{n_B} is the number of Blank Responses.
}
\examples{
## Load example level-2 data
level2 <- invitroTKstats::caco2_L2

\dontrun{
## scenario 1: 
## input level-2 data from the R session and do not export the result table
level3 <- calc_caco2_point(data.in = level2, output.res = FALSE)
}

## scenario 2: 
## import level-2 data from a 'tsv' file and export the result table
\dontrun{
## Refer to sample_verification help file for how to export level-2 data to a directory.
## Unless a different path is specified in OUTPUT.DIR,
## the result table will be saved to the directory specified in INPUT.DIR.
level3 <- calc_caco2_point(FILENAME="Examples", 
                           INPUT.DIR = "invitroTKstats/vignettes")
}

}
\references{
\insertRef{hubatsch2007determination}{invitroTKstats}
}
\author{
John Wambaugh
}
