% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint.R
\name{calc_clint}
\alias{calc_clint}
\title{Calculate intrinsic hepatic clearance}
\usage{
calc_clint(FILENAME, good.col = "Verified")
}
\arguments{
\item{FILENAME}{A string used to identify the input file, whatever the
argument given, "-PPB-RED-Level2.tsv" is appended (defaults to "MYDATA")}

\item{good.col}{Name of a column indicating which rows have been verified for 
analysis, indicated by a "Y" (Defaults to "Verified")}
}
\value{
\item{data.frame}{A data.frame in standardized format}
}
\description{
This function use describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical clearance
as characterized by disappearance of parent compound over time when incubated
with primary hepatocytes (Shibata et al., 2000)
}
\details{
Data are read from a "Level2" text file that should have been formatted and created 
by \code{\link{format_fup_red}} (this is the "Level1" file). The Level1 file
should have been curated and had a column added with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).

The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank \tab Blank\cr
  Hepatocyte inciubation concentration \tab Conc\cr
}

Clint is calculated using \code{\link{lm}} to perform a linear regression of
MS response as a function of time.
}
\examples{

library(invitroTKstats)

clint <- wambaugh2019.clint
clint$Date <- "2019"
clint$Sample.Type <- "Blank"
clint$Time..mins. <- as.numeric(clint$Time..mins.)
clint[!is.na(clint$Time..mins.),"Sample.Type"] <- "Cvst"
clint$ISTD.Name <- "Bucetin, Propranolol, and Diclofenac"
clint$ISTD.Conc <- 1
clint$Dilution.Factor <- 1
clint[is.na(clint$FileName),"FileName"]<-"Wambaugh2019"
clint$Hep.Density <- 0.5
clint$Analysis.Method <- "LC or GC" 
clint$Analysis.Instrument <- "No Idea"
clint$Analysis.Parameters <- "None"

level1 <- format_clint(clint,
  FILENAME="Wambaugh2019",
  sample.col="Sample.Name",
  compound.col="Preferred.Name",
  lab.compound.col="Name",
  time.col="Time..mins.",
  cal.col="FileName")

level2 <- level1
level2$Verified <- "Y"

# All data (allows test for saturation):
write.table(level2,
  file="Wambaugh2019-Clint-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level4 <- calc_clint_point(FILENAME="Wambaugh2019")

}
\author{
John Wambaugh
}
