% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint.R
\name{calc_clint}
\alias{calc_clint}
\title{Calculate Intrinsic Hepatic Clearance (Clint) with Bayesian Modeling (Level-4)}
\usage{
calc_clint(
  FILENAME,
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  good.col = "Verified",
  JAGS.PATH = NA,
  decrease.prob = 0.5,
  saturate.prob = 0.25,
  degrade.prob = 0.05
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the input Level-2 file.
"-Clint-Level2.tsv".}

\item{TEMP.DIR}{(Character) Alternative directory to save output files. 
If \code{NULL}, all files will be written to the current working directory. 
(Defaults to \code{NULL}.)}

\item{NUM.CHAINS}{(Numeric) The number of Markov Chains to use. (Defaults to 5.)}

\item{NUM.CORES}{(Numeric) The number of processors to use for parallel computing. (Defaults to 2.)}

\item{RANDOM.SEED}{(Numeric) The seed used by the random number generator.
(Defaults to 1111.)}

\item{good.col}{(Character) Column name indicating which rows have been 
verified for analysis, valid data rows are indicated with "Y". (Defaults to "Verified".)}

\item{JAGS.PATH}{(Character) Computer specific file path to JAGS software. 
(Defaults to \code{NA}.)}

\item{decrease.prob}{(Numeric) Prior probability that a chemical will decrease in
the assay. (Defaults to 0.5.)}

\item{saturate.prob}{(Numeric) Prior probability that a chemicals rate of metabolism
will decrease between 1 and 10 uM. (Defaults to 0.25.)}

\item{degrade.prob}{(Numeric) Prior probability that a chemical will be unstable
(that is, degrade abiotically) in the assay. (defaults to 0.05.)}
}
\value{
A list of two objects: 
\enumerate{
   \item{Results: A Level-4 data frame with the Bayesian estimated intrinsic hepatic clearance (Clint)
   for 1 and 10 uM and credible intervals for all compounds in the input file. Column includes:
   Compound.Name - compound name, Lab.Compound.Name - compound name used by 
   the laboratory, DTXSID - EPA's DSSTox Structure ID, Clint.1.Med/Clint.10.Med - posterior median, 
   Clint.1.Low/Clint.10.Low - 2.5th quantile, Clint.1.High/Clint.10.High - 97.5th quantile, 
   Clint.pValue, Sat.pValue, degrades.pValue - "p-values" estimated from the probabilities of 
   observing decreases, saturations, and abiotic degradations in all posterior samples.}
   \item{coda: A runjags-class object containing results from JAGS model.}
}
}
\description{
This function estimates the intrinsic hepatic clearance (Clint) with Bayesian
modeling on Hepatocyte Incubation data. Clint at both 1 and 10 uM (if tested) 
and the credible intervals are estimated from posterior samples of the MCMC.
A summary table along with the full set of MCMC results is returned from
the function.
}
\details{
The input to this function should be "Level-2" data. Level-2 data is Level-1,
data formatted with the \code{\link{format_clint}} function, and curated
with a verification column. "Y" in the verification column indicates the
data row is valid for analysis. 

Note: By default, this function writes files to the user's current working
directory. Users must specify an alternative path with the `TEMP.DIR`
argument if they want the files exported to another path. Exported files 
include the summary table (.RData), JAGS model (.RData), and any "unverified" 
data excluded from the analysis (.tsv).

The data frame of observations should be annotated according to
of these types:
\tabular{rl}{
  Blank \tab Cell free blank with media\cr
  CC \tab Cell and media free calibration curve \cr
  Cvst \tab Hepatocyte incubation concentration vs. time \cr
  Inactive \tab Concentration vs. time data with inactivated hepatocytes \cr
}

Clint is calculated using \code{\link{lm}} to perform a linear regression of
MS response as a function of time.
}
\examples{

library(invitroTKstats)

clint <- wambaugh2019.clint
clint$Date <- "2019"
clint$Sample.Type <- "Blank"
clint$Time..mins. <- as.numeric(clint$Time..mins.)
clint[!is.na(clint$Time..mins.),"Sample.Type"] <- "Cvst"
clint$ISTD.Name <- "Bucetin, Propranolol, and Diclofenac"
clint$ISTD.Conc <- 1
clint$Dilution.Factor <- 1
clint[is.na(clint$FileName),"FileName"]<-"Wambaugh2019"
clint$Hep.Density <- 0.5
clint$Analysis.Method <- "LC or GC"
clint$Analysis.Instrument <- "No Idea"
clint$Analysis.Parameters <- "None"

level1 <- format_clint(clint,
  FILENAME="Wambaugh2019",
  sample.col="Sample.Name",
  compound.col="Preferred.Name",
  lab.compound.col="Name",
  time.col="Time..mins.",
  cal.col="FileName")

level2 <- level1
level2$Verified <- "Y"

# All data (allows test for saturation):
write.table(level2,
  file="Wambaugh2019-Clint-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level4 <- calc_clint_point(FILENAME="Wambaugh2019")

}
\author{
John Wambaugh
}
