% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clint_point.R
\name{calc_clint_point}
\alias{calc_clint_point}
\title{Calculate a Point Estimate of Intrinsic Hepatic Clearance (Clint) (Level-3)}
\usage{
calc_clint_point(
  FILENAME,
  good.col = "Verified",
  output.res = TRUE,
  TEMP.DIR = NULL
)
}
\arguments{
\item{FILENAME}{A string used to identify the input Level-2 file.
"<FILENAME>-Clint-Level2.tsv".}

\item{good.col}{(Character) Column name indicating which rows have been
verified, data rows valid for analysis are indicated with a "Y".
(Defaults to "Verified".)}

\item{output.res}{(Logical) When set to \code{TRUE}, the result 
table (Level-3) will be exported the current directory as a .tsv file. 
(Defaults to \code{TRUE}.)}

\item{TEMP.DIR}{(Character) Alternative directory to save output files. By
default, i.e. unspecified, all files will be exported to the user's current
working directory. (Defaults to \code{NULL}.)}
}
\value{
A Level-3 data frame with one row per chemical, containing a point estimate of intrinsic 
clearance (Clint), estimates of Clint of assays performed at 1 and 10 uM (if tested), 
the p-value and the Akaike Information Criterion (AIC) of the linear 
regression fit for all chemicals in the input data frame.
}
\description{
This function calculates a point estimate of intrinsic hepatic clearance (Clint) 
using mass spectrometry (MS) peak area data collected as part of in vitro measurement 
of chemical clearance, as characterized by the disappearance of parent compound over 
time when incubated with primary hepatocytes \insertCite{shibata2002prediction}{invitroTKstats}.
}
\details{
The input to this function should be "Level-2" data. Level-2 data is Level-1,
data formatted with the \code{\link{format_clint}} function, and curated
with a verification column. "Y" in the verification column indicates the
data row is valid for analysis. 

The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank \tab Blank\cr
  Hepatocyte incubation concentration vs. time \tab Cvst\cr
}

Clint is calculated using \code{\link{lm}} to perform a linear regression of
MS response as a function of time.
}
\examples{

library(invitroTKstats)

clint <- wambaugh2019.clint
clint$Date <- "2019"
clint$Sample.Type <- "Blank"
clint$Time..mins. <- as.numeric(clint$Time..mins.)
clint[!is.na(clint$Time..mins.),"Sample.Type"] <- "Cvst"
clint$ISTD.Name <- "Bucetin, Propranolol, and Diclofenac"
clint$ISTD.Conc <- 1
clint$Dilution.Factor <- 1
clint[is.na(clint$FileName),"FileName"]<-"Wambaugh2019"
clint$Hep.Density <- 0.5
clint$Analysis.Method <- "LC or GC"
clint$Analysis.Instrument <- "No Idea"
clint$Analysis.Parameters <- "None"

level1 <- format_clint(clint,
  FILENAME="Wambaugh2019",
  sample.col="Sample.Name",
  compound.col="Preferred.Name",
  lab.compound.col="Name",
  time.col="Time..mins.",
  cal.col="FileName")

level2 <- level1
level2$Verified <- "Y"

# All data (allows test for saturation):
write.table(level2,
  file="Wambaugh2019-Clint-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level3 <- calc_clint_point(FILENAME="Wambaugh2019")

# Just 1 uM data:
write.table(subset(level2,Conc==1),
  file="Wambaugh2019-1-Clint-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level3.1 <- calc_clint_point(FILENAME="Wambaugh2019-1")

# Just 10 uM data:
write.table(subset(level2,Conc==10),
  file="Wambaugh2019-10-Clint-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level3.10 <- calc_clint_point(FILENAME="Wambaugh2019-10")

}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}
}
\author{
John Wambaugh
}
