% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_red.R
\name{calc_fup_red}
\alias{calc_fup_red}
\title{Calculate Fraction Unbound in Plasma (Fup) From Rapid Equilibrium Dialysis (RED) Data}
\usage{
calc_fup_red(
  FILENAME,
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  good.col = "Verified",
  JAGS.PATH = NA,
  Physiological.Protein.Conc = 70/(66.5 * 1000) * 1e+06
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the input Level-2 file.
"<FILENAME>-fup-RED-Level2.tsv".}

\item{TEMP.DIR}{(Character) Alternative directory to save output files. 
If NULL, all files will be written to the current working directory. 
(Defaults to NULL)}

\item{NUM.CHAINS}{(Numeric) The number of Markov Chains to use. (Defaults to 5)}

\item{NUM.CORES}{(Numeric) The number of processors to use. (Defaults to 2)}

\item{RANDOM.SEED}{The seed used by the random number generator.
(Defaults to 1111)}

\item{good.col}{(Character) Column name indicating which rows have been 
verified for analysis, valid data rows indicated with "Y". (Defaults to "Verified")}

\item{JAGS.PATH}{(Character) Computer specific file path to JAGS software. (Defaults to NA)}

\item{Physiological.Protein.Conc}{(Numeric) The assumed physiological protein concentration 
for plasma protein binding calculations. (Defaults to 70/(66.5*1000)*1000000 
according to Berg and Lane (2011): 60-80 mg/mL, albumin is 66.5 kDa, pretend all protein is albumin to get uM.)}
}
\value{
A list of two objects: 
\enumerate{
   \item{A data frame with Bayesian estimated fraction unbound in plasma (Fup) 
   and credible intervals for all compounds in the input file. Column includes:
   Compound.Name - compound name, Lab.Compound.Name - compound name used by 
   the laboratory, DTXSID - EPA's DSSTox Structure ID, Fup.point - point estimate of Fup,
   Fup.Med - posterior median, Fup.Low - 2.5th quantile, and Fup.High - 97.5th quantile}
   \item{A runjags-class object containing results from JAGS model.}
}
}
\description{
This function uses Bayesian modeling, via MCMC simulations, of data collected 
from Rapid Equilibrium Dialysis (RED) assays to estimate the fraction unbound in plasma (Fup) 
and credible interval. This function calculates Fup and the credible interval 
from posterior samples of the MCMC and returns a summary table along with 
the full MCMC results.
}
\details{
The input data should be a "Level-2". Level-2 data is Level-1 data formatted 
with the \code{\link{format_fup_red}} function, then curated with a verification column with 
"Y" indicating the row of data is valid for analysis. 

Note: By default, this function writes files to the user's current working directory 
unless another path is specified in the TEMP.DIR argument. Saved files include 
the summary table (.RData), JAGS model (.RData), and any "unverified" data 
excluded from the analysis (.tsv).

The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  No Plasma Blank (no chemical, no plasma) \tab NoPlasma.Blank\cr
  Plasma Blank (no chemical, just plasma) \tab Plasma.Blank\cr
  Time zero chemical and plasma \tab T0\cr
  Equilibrium chemical in phosphate-buffered well (no plasma) \tab PBS\cr
  Equilibrium chemical in plasma well \tab Plasma\cr
}
}
\examples{
# Level-2 file
write.table(smeltz2023.red,
  file="SmeltzPFAS-fup-RED-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

# JAGS.PATH should be changed to user's specific computer file path to JAGS software.
level4 <- calc_fup_red(FILENAME="SmeltzPFAS",
                       NUM.CORES=8,
                       JAGS.PATH="C:/Users/jwambaug/AppData/Local/JAGS/JAGS-4.3.0/x64")

}
\references{
\insertRef{waters2008validation}{invitroTKstats}

\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\author{
John Wambaugh and Chantel Nicolas
}
