% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_red.R
\name{calc_fup_red}
\alias{calc_fup_red}
\title{Calculate Fraction Unbound in Plasma From Rapid Equilibrium Dialysis Data}
\usage{
calc_fup_red(
  FILENAME,
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  good.col = "Verified",
  JAGS.PATH = NA,
  Physiological.Protein.Conc = 70/(66.5 * 1000) * 1e+06
)
}
\arguments{
\item{FILENAME}{A string used to identify the input Level-2 file, whatever the
argument given, "-fup-RED-Level2.tsv" is appended.}

\item{TEMP.DIR}{Alternative directory to output file
(Defaults to NULL, all files will be written to user's current working directory).}

\item{NUM.CHAINS}{The number of Markov Chains to use (Defaults to 5).}

\item{NUM.CORES}{The number of processors to use (Defaults to 2).}

\item{RANDOM.SEED}{The seed used by the random number generator
(Defaults to 1111).}

\item{good.col}{Name of a column indicating which rows have been verified for
analysis, indicated by a "Y" (Defaults to "Verified").}

\item{JAGS.PATH}{The file path to JAGS (Defaults to NULL).}

\item{Physiological.Protein.Conc}{Defaults to 70/(66.5*1000)*1000000 per Berg and Lane (2011), 
60-80 mg/mL, albumin is 66.5 kDa, pretend all protein is albumin to get uM.}
}
\value{
A list of two objects: 
\enumerate{
   \item{A data frame containing quantiles of the Bayesian posteriors of 
   fraction unbound in plasma (Fup) of all compounds in input file. Column includes:
   Compound.Name - compound name, Lab.Compound.Name - compound name used by 
   the laboratory, DTXSID - EPA's DSSTox Structure ID, Fup.point - point estimate of Fup,
   Fup.Med - median of posteriors, Fup.Low - 2.5th quantile, and Fup.High - 97.5th quantile}
   \item{A runjags-class object containing results from JAGS model.}
}
}
\description{
This function uses MCMC simulation to calculate fraction unbound in plasma (Fup) 
from rapid equilibrium dialysis (RED) data. This function then calculates quantiles 
of posterior samples and returns a summary table along with all MCMC results.
}
\details{
The input data should be a "Level-2" data that have been formatted and created
by \code{\link{format_fup_red}}, then have been curated and had a column added
with the value "Y" indicating that each row is verified as usable for analysis. 

Be aware that this function will write couple files to user's current working directory
unless other path is specified in the argument. Files being saved include the 
summary table (.RData), JAG argument (.RData), and any "unverified" data that 
were held out from the analysis (.tsv).  

The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  No Plasma Blank (no chemical, no plasma) \tab NoPlasma.Blank\cr
  Plasma Blank (no chemical, just plasma) \tab Plasma.Blank\cr
  Time zero chemical and plasma \tab T0\cr
  Equilibrium chemical in phosphate-buffered well (no plasma) \tab PBS\cr
  Equilibrium chemical in plasma well \tab Plasma\cr
}
}
\examples{
# Level-2 file
write.table(smeltz2023.red,
  file="SmeltzPFAS-fup-RED-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level4 <- calc_fup_red(FILENAME="SmeltzPFAS",
                       NUM.CORES=8,
                       JAGS.PATH="C:/Users/jwambaug/AppData/Local/JAGS/JAGS-4.3.0/x64")

}
\references{
\insertRef{waters2008validation}{invitroTKstats}

\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\author{
John Wambaugh and Chantel Nicolas
}
