% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_red_fup_base.R
\name{calc_fup_red_base}
\alias{calc_fup_red_base}
\title{Calculate fraction unbound in plasma from rapid equilibruim dialysis data}
\usage{
calc_fup_red_base(
  PPB.data,
  this.conc = 100,
  FILENAME = "BASE_Model_Results",
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  compound.col = "Compound.Name",
  area.col = "Area",
  series.col = "Series",
  type.col = "Sample.Type",
  compound.conc.col = "Nominal.Conc",
  cal.col = "Cal",
  dilution.col = "Dilution.Factor",
  istd.col = "ISTD.Area",
  istd.name.col = "ISTD.Name",
  istd.conc.col = "ISTD.Conc",
  nominal.test.conc.col = "Test.Target.Conc"
)
}
\arguments{
\item{PPB.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type.}

\item{this.conc}{The plasma protein concentration relative to physiologic
levels (default 100%)}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "BASE_Model_Results")}

\item{TEMP.DIR}{An optional directory where file writing may be faster.}

\item{NUM.CHAINS}{The number of Markov Chains to use. This allows evaluation
of convergence according to Gelman and Rubin diagnostic.}

\item{NUM.CORES}{The number of processors to use (default 2)}

\item{JAGS.PATH}{The file path to JAGS.}

\item{RANDOM.SEED}{The seed used by the random number generator 
(default 1111)}
}
\value{
A data.frame containing quunantiles of the Bayesian posteriors
}
\description{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Sample Blank (no chemical, just plasma) \tab Blank\cr
  Time zero chemical and plasma \tab T0\cr
  Equilibrium chemical in phosphate-buffered well (no plasma) \tab PBS\cr
  Equilibrium chemical in plasma well \tab Plasma\cr
}
}
\author{
John Wambaugh and Chantel Nicolas
}
