% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_red_point.R, R/calc_fup_uc_point.R
\name{calc_fup_red_point}
\alias{calc_fup_red_point}
\title{Calculate a point estimate of Fraction Unbound in Plasma (RED)}
\usage{
calc_fup_red_point(FILENAME, good.col = "Verified")

calc_fup_red_point(FILENAME, good.col = "Verified")
}
\arguments{
\item{FILENAME}{A string used to identify the input file, whatever the
argument given, "-PPB-RED-Level2.tsv" is appended (defaults to "MYDATA")}

\item{good.col}{Name of a column indicating which rows have been verified for 
analysis, indicated by a "Y" (Defaults to "Verified")}
}
\value{
\item{data.frame}{A data.frame in standardized format}

\item{data.frame}{A data.frame in standardized format}
}
\description{
This function use describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using rapid equilibrium dialysis (Waters, et al, 2008).
Data are read from a "Level2" text file that should have been formatted and created 
by \code{\link{format_fup_red}} (this is the "Level1" file). The Level1 file
should have been curated and had a column added with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).

This function use describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using ultracentrifugation (Redgrave 1975?).
Data are read from a "Level2" text file that should have been formatted and created 
by \code{\link{format_fup_red}} (this is the "Level1" file). The Level1 file
should have been curated and had a column added with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).
}
\details{
The should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}

F_up is calculated from MS responses as:

f_up = mean(PBS Response * Dilution.Factor) / mean(Plasma Response * Dilution Factor)

The should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}

F_up is calculated from MS responses as:

f_up = mean(AF Response * Dilution.Factor) / mean(T5 Response * Dilution Factor)
}
\references{
Waters, Nigel J., et al. "Validation of a rapid equilibrium dialysis 
approach for the measurement of plasma protein binding." Journal of 
Pharmaceutical Sciences 97.10 (2008): 4586-4595.

Redgrave, T. G., D. C. K. Roberts, and C. E. West. "Separation of plasma 
lipoproteins by density-gradient ultracentrifugation." Analytical 
Biochemistry 65.1-2 (1975): 42-49.#'
}
\author{
John Wambaugh

John Wambaugh
}
