% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_red_point.R
\name{calc_fup_red_point}
\alias{calc_fup_red_point}
\title{Calculate a point estimate of Fraction Unbound in Plasma (RED)}
\usage{
calc_fup_red_point(FILENAME, good.col = "Verified")
}
\arguments{
\item{FILENAME}{A string used to identify the input file, whatever the
argument given, "-fup-RED-Level2.tsv" is appended (defaults to "MYDATA")}

\item{good.col}{Name of a column indicating which rows have been verified for
analysis, indicated by a "Y" (Defaults to "Verified")}
}
\value{
\item{data.frame}{A data.frame in standardized format}
}
\description{
This function uses mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using rapid equilibrium dialysis
\insertCite{waters2008validation}{invitroTKstats}.
Data are read from a "Level2" text file that should have been formatted and created
by \code{\link{format_fup_red}} (this is the "Level1" file). The Level1 file
should have been curated and had a column added with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  No Plasma Blank (no chemical, no plasma) \tab NoPlasma.Blank\cr
  Plasma Blank (no chemical, just plasma) \tab Plasma.Blank\cr
  Time zero chemical and plasma \tab T0\cr
  Equilibrium chemical in phosphate-buffered well (no plasma) \tab PBS\cr
  Equilibrium chemical in plasma well \tab Plasma\cr
}

F_up is calculated from MS responses as:

f_up = max(0,(mean(PBS Response * Dilution.Factor) -
  mean(NoPlasma.Blank Response * Dilution.Factor))) / (
  mean(Plasma Response * Dilution Factor) -
  mean(Plasma.Blank Response * Dilution.Factor))
}
\examples{
red <- subset(wambaugh2019.red, Protein==100)
red$Date <- "2019"
red$Sample.Type <- "Blank"
red <- subset(red,!is.na(SampleName))
red[regexpr("PBS",red$SampleName)!=-1,"Sample.Type"] <- "PBS"
red[regexpr("Plasma",red$SampleName)!=-1,"Sample.Type"] <- "Plasma"
red$Dilution.Factor <- NA
red$Dilution.Factor <- as.numeric(red$Dilution.Factor)
red[red$Sample.Type=="PBS","Dilution.Factor"] <- 2
red[red$Sample.Type=="Plasma","Dilution.Factor"] <- 5
red[regexpr("T0",red$SampleName)!=-1,"Sample.Type"] <- "T0"
red$Analysis.Method <- "LC or GC"
red$Analysis.Instrument <- "No Idea"
red$Analysis.Parameters <- "None"


# Strip out protein conc information from compound names:
red$CompoundName <- gsub("-100P","",red$CompoundName)
red$CompoundName <- gsub("-30P","",red$CompoundName)
red$CompoundName <- gsub("-10P","",red$CompoundName)

red$Test.Target.Conc <- 5
red$ISTD.Name <- "Bucetin and Diclofenac"
red$ISTD.Conc <- 1
red$Series <- 1

level1 <- format_fup_red(red,
  FILENAME="Wambaugh2019",
  sample.col="SampleName",
  compound.col="Preferred.Name",
  lab.compound.col="CompoundName",
  cal.col="RawDataSet")

level2 <- level1
level2$Verified <- "Y"

write.table(level2,
  file="Wambaugh2019-fup-RED-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level3 <- calc_fup_red_point(FILENAME="Wambaugh2019")

}
\references{
\insertRef{waters2008validation}{invitroTKstats}
}
\author{
John Wambaugh
}
