% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_uc.R
\name{calc_fup_uc}
\alias{calc_fup_uc}
\title{Calculate Fraction Unbound in Plasma (Fup) from Ultracentrifugation data}
\usage{
calc_fup_uc(
  FILENAME = "UC_Model_Results",
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  good.col = "Verified",
  JAGS.PATH = NA
)
}
\arguments{
\item{FILENAME}{A string used to identify the input Level-2 file, whatever the
argument given, "-fup-UC-Level2.tsv" is appended.}

\item{TEMP.DIR}{Alternative directory to output file
(Defaults to NULL, all files will be written to user's current working directory).}

\item{NUM.CHAINS}{The number of Markov Chains to use (Defaults to 5).}

\item{NUM.CORES}{The number of processors to use (Defaults to 2).}

\item{RANDOM.SEED}{The seed used by the random number generator
(Defaults to 1111).}

\item{good.col}{Name of a column indicating which rows have been verified for
analysis, indicated by a "Y" (Defaults to "Verified").}

\item{JAGS.PATH}{The file path to JAGS (Defaults to NA).}
}
\value{
A list of two objects: 
\enumerate{
   \item{A data frame containing quantiles of the Bayesian posteriors of 
   fraction unbound in plasma (Fup) of all compounds in input file. Column includes:
   Compound.Name - compound name, Lab.Compound.Name - compound name used by 
   the laboratory, DTXSID - EPA's DSSTox Structure ID, Fup.point - point estimate of Fup,
   Fup.Med - median of posteriors, Fup.Low - 2.5th quantile, and Fup.High - 97.5th quantile}
   \item{A runjags-class object containing results from JAGS model.}
}
}
\description{
This function uses MCMC simulation to calculate fraction unbound in plasma (Fup) 
from Ultracentrifugation (UC) data. This function then calculates quantiles 
of posterior samples and returns a summary table along with all MCMC results.
}
\details{
The input data should be a "Level-2" data that have been formatted and created
by \code{\link{format_fup_uc}}, then have been curated and had a column added
with the value "Y" indicating that each row is verified as usable for analysis. 

Be aware that this function will write couple files to user's current working directory
unless other path is specified in the argument. Files being saved include the 
summary table (.RData), JAG argument (.RData), and any "unverified" data that 
were held out from the analysis (.tsv).  

The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultracentrifugation Aqueous Fraction \tab UC\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}
We don't currently use the T1 data, but you must have CC, AF, and T5 data.
}
\author{
John Wambaugh and Chantel Nicolas
}
