% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_uc.R
\name{calc_fup_uc}
\alias{calc_fup_uc}
\title{Calculate fraction unbound in plasma from ultracentrifugation data}
\usage{
calc_fup_uc(
  PPB.data,
  FILENAME = "UC_Model_Results",
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  good.col = "Verified",
  JAGS.PATH = NA
)
}
\arguments{
\item{PPB.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "BASE_Model_Results")}

\item{TEMP.DIR}{An optional directory where file writing may be faster.}

\item{NUM.CHAINS}{The number of Markov Chains to use. This allows evaluation
of convergence according to Gelman and Rubin diagnostic.}

\item{NUM.CORES}{The number of processors to use (default 2)}

\item{RANDOM.SEED}{The seed used by the random number generator 
(default 1111)}

\item{JAGS.PATH}{The file path to JAGS.}

\item{sample.col}{Which column of PPB.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{lab.compound.col}{Which column of PPB.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{dtxsid.col}{Which column of PPB.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{date.col}{Which column of PPB.data indicates the laboratory measurement
date (Defaults to "Date")}

\item{compound.col}{Which column of PPB.data indicates the test compound
(Defaults to "Compound.Name")}

\item{area.col}{Which column of PPB.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate (Defaults to "Series")}

\item{type.col}{Which column of PPB.data indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{cal.col}{Which column of PPB.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")

#param std.conc.col Which column indicates the intended concentration 
of the test chemical for calibration curves (Defaults to "Standard.Conc")}

\item{dilution.col}{Which column of PPB.data indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{istd.col}{Which column of PPB.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{istd.name.col}{Which column of PPB.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc.col}{Which column of PPB.data indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{uc.assay.conc.col}{Which column of PPB.data indicates the intended
test chemical concentration at time zero (Defaults to "UC.Assay.Conc")}
}
\value{
A data.frame containing quunantiles of the Bayesian posteriors
}
\description{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultracentrifugation Aqueous Fraction \tab UC\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}
We don't currently use the T1 data, but you must have CC, AF, and T5 data.
}
\author{
John Wambaugh and Chantel Nicolas
}
