% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_uc_fup.R
\name{calc_fup_uc}
\alias{calc_fup_uc}
\title{Calculate fraction unbound in plasma from ultracentrifugation data}
\usage{
calc_fup_uc(
  PPB.data,
  FILENAME = "UC_Model_Results",
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  compound.col = "Compound.Name",
  area.col = "Area",
  series.col = "Series",
  type.col = "Sample.Type",
  compound.conc.col = "Nominal.Conc",
  cal.col = "Cal",
  dilution.col = "Dilution.Factor",
  istd.col = "ISTD.Area",
  istd.name.col = "ISTD.Name",
  istd.conc.col = "ISTD.Conc",
  nominal.test.conc.col = "Test.Target.Conc"
)
}
\arguments{
\item{PPB.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type.}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "BASE_Model_Results")}

\item{TEMP.DIR}{An optional directory where file writing may be faster.}

\item{NUM.CHAINS}{The number of Markov Chains to use. This allows evaluation
of convergence according to Gelman and Rubin diagnostic.}

\item{NUM.CORES}{The number of processors to use (default 2)}

\item{RANDOM.SEED}{The seed used by the random number generator 
(default 1111)}

\item{this.conc}{The plasma protein concentration relative to physiologic
levels (default 100%)}

\item{JAGS.PATH}{The file path to JAGS.}
}
\value{
\item{data.frame}{A data.frame containing quunantiles of the 
Bayesian posteriors}
}
\description{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultra-centrifugation Aqueous Fraction \tab UC\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}
We don't currently use the T1 data, but you must have CC, AF, and T5 data.
}
\author{
John Wambaugh and Chantel Nicolas
}
