% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_uc.R
\name{calc_fup_uc}
\alias{calc_fup_uc}
\title{Calculate the Fraction Unbound in Plasma (Fup) from Ultracentrifugation (UC) Data}
\usage{
calc_fup_uc(
  FILENAME,
  data.in,
  TEMP.DIR = NULL,
  NUM.CHAINS = 5,
  NUM.CORES = 2,
  RANDOM.SEED = 1111,
  good.col = "Verified",
  JAGS.PATH = NA,
  save.MCMC = FALSE,
  sig.figs = 3,
  INPUT.DIR = NULL,
  OUTPUT.DIR = NULL
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the input Level-2 file,
"<FILENAME>-fup-UC-Level2.tsv", and to name the exported model results. 
This argument is required no matter which method of specifying input data is used. 
(Defaults to \code{NULL}.)}

\item{data.in}{A Level-2 data frame generated from the 
\code{format_fup_uc} function with a verification column added by 
\code{sample_verification}. Complement with manual verification if needed.}

\item{TEMP.DIR}{(Character) Temporary directory to save intermediate files. By
default, i.e. unspecified, all files will be exported to the user's current
working directory. (Defaults to \code{NULL}.)}

\item{NUM.CHAINS}{(Numeric) The number of Markov Chains to use. (Defaults to 5.)}

\item{NUM.CORES}{(Numeric) The number of computer processors to use for
parallel computing. (Defaults to 2.)}

\item{RANDOM.SEED}{(Numeric) The seed used by the random number generator.
(Defaults to 1111.)}

\item{good.col}{(Character) Column name indicating which rows have been
verified, data rows valid for analysis are indicated with a "Y".
(Defaults to "Verified".)}

\item{JAGS.PATH}{(Character) Computer specific file path to JAGS software.
(Defaults to `NA`.)}

\item{save.MCMC}{(Logical) When set to \code{TRUE}, will export the MCMC results
as an .RData file. (Defaults to \code{FALSE}.)}

\item{sig.figs}{(Numeric) The number of significant figures to round the exported unverified data (Level-2). 
The exported result table (Level-4) is left unrounded for reproducibility.
(Note: console print statements are also rounded to specified significant figures.)
(Defaults to \code{3}.)}

\item{INPUT.DIR}{(Character) Path to the directory where the input level-2 file exists. 
If \code{NULL}, looking for the input level-2 file in the current working
directory. (Defaults to \code{NULL}.)}

\item{OUTPUT.DIR}{(Character) Path to the directory to save the output file. 
If \code{NULL}, the output file will be saved to the current working
directory. (Defaults to \code{NULL}.)}
}
\value{
A list of two objects: 
\enumerate{
   \item{Results: A data frame with Bayesian estimated fraction unbound
   in plasma (Fup) and credible intervals for all compounds in the input file.
   Column includes:
   Compound.Name - compound name,
   Lab.Compound.Name - compound name used by the laboratory,
   DTXSID - EPA's DSSTox Structure ID,
   Fup.point - point estimate of Fup,
   Fup.Med - Posterior median,
   Fup.Low - 2.5th quantile,
   and Fup.High - 97.5th quantile.}
   \item{coda: A runjags-class object containing results from JAGS model.}
}
}
\description{
This function estimates the fraction unbound in plasma (Fup) and credible
intervals with a Bayesian modeling approach, via MCMC simulations.
Data used in modeling is collected from Ultracentrifugation (UC) Fup assays.
Fup and the credible interval are calculated from the MCMC posterior samples
and the function returns a summary table along with the full set of
MCMC results.
}
\details{
The input to this function should be "Level-2" data. Level-2 data is Level-1,
data formatted with the \code{\link{format_fup_uc}} function, and curated
with a verification column. "Y" in the verification column indicates the
data row is valid for analysis. 

Note: By default, this function writes files to the user's current working
directory. Users must specify an alternative path with the `TEMP.DIR`
argument if they want the files exported to another path. Exported files 
include the summary table (.RData), JAGS model (.RData), and any "unverified" 
data excluded from the analysis (.tsv).

The data frame of observations should be annotated according to
these types:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultracentrifugation Aqueous Fraction \tab AF\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}
We don't currently use the T1 data, but CC, AF, and T5 data are required.
}
\examples{
## Example 1: loading level-2 using data.in
\dontrun{
level2 <- invitroTKstats::fup_uc_L2

# JAGS.PATH should be changed to user's specific computer file path to JAGS software.
# findJAGS() from runjags package is a handy function to find JAGS path automatically.
# In certain circumstances or cases, one may need to provide the absolute path to JAGS.
path.to.JAGS <- findJAGS()
level4 <- calc_fup_uc(data.in = level2,
                      NUM.CORES=2,
                      JAGS.PATH=path.to.JAGS)
}

## Example 2: importing level-2 from a .tsv file
\dontrun{
# Refer to sample_verification help file for how to export level-2 data to a directory.
# JAGS.PATH should be changed to user's specific computer file path to JAGS software.
# findJAGS() from runjags package is a handy function to find JAGS path automatically.
# In certain circumstances or cases, one may need to provide the absolute path to JAGS.
path.to.JAGS <- findJAGS()
level4 <- calc_fup_uc(FILENAME="KreutzPFAS",
                      NUM.CORES=2,
                      JAGS.PATH=path.to.JAGS,
                      INPUT.DIR = "invitroTKstats/vignettes")
}

}
\author{
John Wambaugh and Chantel Nicolas
}
