% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_uc_point.R
\name{calc_fup_uc_point}
\alias{calc_fup_uc_point}
\title{Calculate Point Estimates of Fraction Unbound in Plasma with
Ultracentrifugation (UC) Data}
\usage{
calc_fup_uc_point(FILENAME, good.col = "Verified")
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the input Level-2 file.
"<FILENAME>-fup-UC-Level2.tsv".}

\item{good.col}{(Character) Column name indicating which rows have been
verified, data rows valid for analysis are indicated with a "Y".
(Defaults to "Verified".)}
}
\value{
A data frame with one row per chemical, contains chemical identifiers 
such as preferred compound name, compound name used by the laboratory, 
EPA's DSSTox Structure ID, calibration, and point estimates for
the fraction unbound in plasma (Fup) for all chemicals in the input data frame.
}
\description{
This function calculates the point estimates for the fraction unbound in
plasma (Fup) using mass spectrometry (MS) peak areas from samples collected
as part of in vitro measurements of chemical Fup using ultracentrifugation
\insertCite{waters2008validation}{invitroTKstats}. See the Details section
for the equation(s) used in the point estimate.
}
\details{
The input to this function should be "Level-2" data. Level-2 data is Level-1,
data formatted with the \code{\link{format_fup_uc}} function, and curated
with a verification column. "Y" in the verification column indicates the
data row is valid for analysis. 

The should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}

\eqn{f_{up}} is calculated from MS responses as:

\eqn{f_{up} = \frac{\sum_{i = 1}^{n_A} (r_A * c_{DF}) / n_A}{\sum_{i = 1}^{n_{T5}} (r_{T5} * c_{DF}) / n_{T5}}}

where \eqn{r_A} is Aqueous Fraction Response, \eqn{c_{DF}} is Dilution Factor,
\eqn{r_{T5}} is T5 Response, \eqn{n_A} is the number of Aqueous Fraction Responses,
and \eqn{n_{T5}} is the number of T5 Responses.
}
\examples{
level0 <- kreutz2020
level0$Analysis.Method <- "GC"
level0$Analysis.Instrument <- "No Idea"
level0$Analysis.Parameters <- "None"
level1 <- format_fup_uc(level0,
  FILENAME="Kreutz2020",
  compound.col="Name",
  compound.conc.col="Standard.Conc",
  area.col="Chem.Area"
  )
level2 <- level1
level2$Verified <- "Y"

write.table(level2,
  file="Kreutz2020-fup-UC-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level3 <- calc_fup_uc_point(FILENAME="Kreutz2020")

}
\references{
\insertRef{redgrave1975separation}{invitroTKstats}
}
\author{
John Wambaugh
}
