% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_uc_point.R
\name{calc_fup_uc_point}
\alias{calc_fup_uc_point}
\title{Calculate a point estimate of Fraction Unbound in Plasma (UC)}
\usage{
calc_fup_uc_point(FILENAME, good.col = "Verified")
}
\arguments{
\item{FILENAME}{A string used to identify the input file, whatever the
argument given, "-PPB-UC-Level2.tsv" is appended (defaults to "MYDATA")}

\item{good.col}{Name of a column indicating which rows have been verified for
analysis, indicated by a "Y" (Defaults to "Verified")}
}
\value{
\item{data.frame}{A data.frame in standardized format}
}
\description{
This function use describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using ultracentrifugation
\insertCite{redgrave1975separation}{invitroTKstats}.
Data are read from a "Level2" text file that should have been formatted and created
by \code{\link{format_fup_red}} (this is the "Level1" file). The Level1 file
should have been curated and had a column added with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).
}
\details{
The should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}

\eqn{f_{up}} is calculated from MS responses as:

\eqn{f_{up} = \frac{\sum_{i = 1}^n(r_A * c_{DF}) / n_A}{\sum_{i = 1}^n(r_{T5} * c_{DF}) / n_{T5}}}

where \eqn{r_A} is Aqueous Fraction Response, \eqn{c_{DF}} is Dilution Factor,
\eqn{r_{T5}} is T5 Response, \eqn{n_A} is the number of Aqueous Fraction Responses,
and \eqn{n_{T5}} is the number of T5 Responses.
}
\examples{
level0 <- kreutz2020
level0$Analysis.Method <- "GC"
level0$Analysis.Instrument <- "No Idea"
level0$Analysis.Parameters <- "None"
level1 <- format_fup_uc(level0,
  FILENAME="Kreutz2020",
  compound.col="Name",
  compound.conc.col="Standard.Conc",
  area.col="Chem.Area"
  )
level2 <- level1
level2$Verified <- "Y"

write.table(level2,
  file="Kreutz2020-fup-UC-Level2.tsv",
  sep="\t",
  row.names=F,
  quote=F)

level3 <- calc_fup_uc_point(FILENAME="Kreutz2020")

}
\references{
\insertRef{redgrave1975separation}{invitroTKstats}
}
\author{
John Wambaugh
}
