% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fup_uc_point.R
\name{calc_fup_uc_point}
\alias{calc_fup_uc_point}
\title{Calculate Point Estimates of Fraction Unbound in Plasma with
Ultracentrifugation (UC) Data}
\usage{
calc_fup_uc_point(
  FILENAME,
  data.in,
  good.col = "Verified",
  output.res = TRUE,
  sig.figs = 3,
  INPUT.DIR = NULL,
  OUTPUT.DIR = NULL
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the input Level-2 file.
"<FILENAME>-fup-UC-Level2.tsv".}

\item{data.in}{(Data Frame) A Level-2 data frame generated from the 
\code{format_fup_uc} function with a verification column added by 
\code{sample_verification}. Complement with manual verification if needed.}

\item{good.col}{(Character) Column name indicating which rows have been
verified, data rows valid for analysis are indicated with a "Y".
(Defaults to "Verified".)}

\item{output.res}{(Logical) When set to \code{TRUE}, the result 
table (Level-3) will be exported the current directory as a .tsv file. 
(Defaults to \code{TRUE}.)}

\item{sig.figs}{(Numeric) The number of significant figures to round the exported result table (Level-3). 
(Note: console print statements are also rounded to specified significant figures.) 
(Defaults to \code{3}.)}

\item{INPUT.DIR}{(Character) Path to the directory where the input level-2 file exists. 
If \code{NULL}, looking for the input level-2 file in the current working
directory. (Defaults to \code{NULL}.)}

\item{OUTPUT.DIR}{(Character) Path to the directory to save the output file. 
If \code{NULL}, the output file will be saved to the current working
directory or \code{INPUT.DIR} if specified. (Defaults to \code{NULL}.)}
}
\value{
A data frame with one row per chemical, contains chemical identifiers 
such as preferred compound name, compound name used by the laboratory, 
EPA's DSSTox Structure ID, calibration, and point estimates for
the fraction unbound in plasma (Fup) for all chemicals in the input data frame.
}
\description{
This function calculates the point estimates for the fraction unbound in
plasma (Fup) using mass spectrometry (MS) peak areas from samples collected
as part of in vitro measurements of chemical Fup using ultracentrifugation
\insertCite{waters2008validation}{invitroTKstats}. See the Details section
for the equation(s) used in the point estimate.
}
\details{
The input to this function should be "Level-2" data. Level-2 data is Level-1,
data formatted with the \code{\link{format_fup_uc}} function, and curated
with a verification column. "Y" in the verification column indicates the
data row is valid for analysis. 

The should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}

\eqn{f_{up}} is calculated from MS responses as:

\eqn{f_{up} = \frac{\sum_{i = 1}^{n_A} (r_A * c_{DF}) / n_A}{\sum_{i = 1}^{n_{T5}} (r_{T5} * c_{DF}) / n_{T5}}}

where \eqn{r_A} is Aqueous Fraction Response, \eqn{c_{DF}} is Dilution Factor,
\eqn{r_{T5}} is T5 Response, \eqn{n_A} is the number of Aqueous Fraction Responses,
and \eqn{n_{T5}} is the number of T5 Responses.
}
\examples{
## Load example level-2 data
level2 <- invitroTKstats::fup_uc_L2

## scenario 1: 
## input level-2 data from the R session and do not export the result table
level3 <- calc_fup_uc_point(data.in = level2, output.res = FALSE)

## scenario 2: 
## import level-2 data from a 'tsv' file and export the result table
\dontrun{
## Refer to sample_verification help file for how to export level-2 data to a directory.
## Unless a different path is specified in OUTPUT.DIR,
## the result table will be saved to the directory specified in INPUT.DIR.
level3 <- calc_fup_uc_point(FILENAME="KreutzPFAS", 
                            INPUT.DIR = "invitroTKstats/vignettes")
}

}
\references{
\insertRef{redgrave1975separation}{invitroTKstats}
}
\author{
John Wambaugh
}
