% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chem_table.R
\name{create_chem_table}
\alias{create_chem_table}
\title{Creates a standardized data table of chemical identities}
\usage{
create_chem_table(
  input.table,
  dtxsid.col = "DTXSID",
  compound.col = "Compound.Name",
  lab.compound.col = "Lab.Compound.Name"
)
}
\arguments{
\item{input.table}{A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and analytical chemistry methods.
It should contain columns with names specified by the following arguments:}

\item{dtxsid.col}{Which column of input.table indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{compound.col}{Which column of input.table indicates the test compound
(Defaults to "Compound.Name")}

\item{lab.compound.col}{Which column of PPB.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}
}
\value{
A data.frame with one row per method
}
\description{
This function extracts the chemical analysis methods from a set of MS data.
}
\examples{

library(invitroTKstats)

# Smeltz et al. (2020) data:
create_chem_table(smeltz2020)

# Kreutz et al. (2020) data:
create_chem_table(kreutz2020,compound.col="Name")

# Wambaugh et al. (2019) data:
# Strip out protein conc information from compound names:
wambaugh2019.red$CompoundName <- gsub("-100P","",wambaugh2019.red$CompoundName)
wambaugh2019.red$CompoundName <- gsub("-30P","",wambaugh2019.red$CompoundName)
wambaugh2019.red$CompoundName <- gsub("-10P","",wambaugh2019.red$CompoundName)

wambaugh2019.redchems <- create_chem_table(wambaugh2019.red,
   compound.col="Preferred.Name",
   lab.compound.col="CompoundName")

}
\author{
John Wambaugh
}
