% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chem_table.R
\name{create_chem_table}
\alias{create_chem_table}
\title{Creates a Standardized Data Table of Chemical Identities}
\usage{
create_chem_table(
  input.table,
  dtxsid.col = "DTXSID",
  compound.col = "Compound.Name",
  lab.compound.col = "Lab.Compound.Name"
)
}
\arguments{
\item{input.table}{(Data Frame) A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and analytical chemistry methods.
It should contain columns with names specified by the following arguments:}

\item{dtxsid.col}{(Character) Column name of input.table containing EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}). (Defaults to "DTXSID".)}

\item{compound.col}{(Character) Column name of input.table containing the test compound.
(Defaults to "Compound.Name".)}

\item{lab.compound.col}{(Character) Column name of input.table containing the test compound 
name used by the laboratory. (Defaults to "Lab.Compound.Name".)}
}
\value{
A data.frame with one row per chemical containing compound names and laboratory names used
for each chemical in the input data frame.
}
\description{
This function creates a data table that summarizes different names used for each chemical 
from a set of MS data. Each row in the returned data table represents an unique chemical 
defined by EPA's DSSTox Structure ID (dtxsid) and contains all compound names and names 
used by laboratory for that particular chemical in the data.
}
\examples{

library(invitroTKstats)

# Smeltz et al. (2020) data:
create_chem_table(smeltz2020)

# Kreutz et al. (2020) data:
create_chem_table(kreutz2020,compound.col="Name")

}
\author{
John Wambaugh
}
