% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_method_table.R
\name{create_method_table}
\alias{create_method_table}
\title{Creates a standardized data table for chemical analysis methods}
\usage{
create_method_table(
  input.table,
  dtxsid.col = "DTXSID",
  compound.col = "Compound.Name",
  istd.name.col = "ISTD.Name",
  analysis.method.col = "Analysis.Method",
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters.col = "Analysis.Parameters"
)
}
\arguments{
\item{input.table}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and analytical chemistry methods.
It should contain columns with names specified by the following arguments:}

\item{dtxsid.col}{Which column of input.table indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{compound.col}{Which column of input.table indicates the test compound
(Defaults to "Compound.Name")}

\item{istd.name.col}{Which column of PPB.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{analysis.method.col}{Which column of PPB.data indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument.col}{Which column of PPB.data indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters.col}{Which column of PPB.data indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}
}
\value{
A data.frame with one row per method
}
\description{
This function extracts the chemical analysis methods from a set of MS data.
}
\author{
John Wambaugh
}
