% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_method_table.R
\name{create_method_table}
\alias{create_method_table}
\title{Creates a standardized data table for chemical analysis methods}
\usage{
create_method_table(
  input.table,
  dtxsid.col = "DTXSID",
  compound.col = "Compound.Name",
  istd.name.col = "ISTD.Name",
  analysis.method.col = "Analysis.Method",
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters.col = "Analysis.Parameters"
)
}
\arguments{
\item{input.table}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and analytical chemistry methods.
It should contain columns with names specified by the following arguments:}

\item{dtxsid.col}{Which column of input.table indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{compound.col}{Which column of input.table indicates the test compound
(Defaults to "Compound.Name")}

\item{istd.name.col}{Which column of PPB.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{analysis.method.col}{Which column of PPB.data indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument.col}{Which column of PPB.data indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters.col}{Which column of PPB.data indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}
}
\value{
A data.frame with one row per method
}
\description{
This function extracts the chemical analysis methods from a set of MS data.
}
\examples{
library(invitroTKstats)

create_method_table(smeltz2020)

create_method_table(kreutz2020,compound.col="Name")

# Create the method columns invitroTKstats is expecting from the
# Wambaugh et al. (2019) supplemental table 10:
wambaugh2019.methods$Method <- ""
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$LC),"LC"] <- ""
# Describe the chemcials with LC as such:
wambaugh2019.methods[wambaugh2019.methods$LC=="Y","Method"] <- "LC"
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$GC),"GC"] <- ""
# Describe the chemcials with GC as such:
wambaugh2019.methods[wambaugh2019.methods$GC=="Y","Method"] <- "GC"
# Remove the non GC/LC-able chemicals:

# Set instruments and notes:
# Agilent QQQ:
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$Agilent.QQQ),"Agilent.QQQ"] <-
  "Failed"
wambaugh2019.methods[wambaugh2019.methods$Agilent.QQQ!="Failed","Instrument"] <-
  "Agilent QQQ"
wambaugh2019.methods[wambaugh2019.methods$Agilent.QQQ!="Failed",
  "Analysis.Notes"] <- 
  wambaugh2019.methods[wambaugh2019.methods$Agilent.QQQ!="Failed","Agilent.QQQ"]

# Water's Xevo:
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$Water.s.Xevo),"Water.s.Xevo"] <-
  "Failed"
wambaugh2019.methods[wambaugh2019.methods$Water.s.Xevo!="Failed",
  "Instrument"] <- "Waters Xevo"
wambaugh2019.methods[wambaugh2019.methods$Water.s.Xevo!="Failed",
  "Analysis.Notes"] <- 
  wambaugh2019.methods[wambaugh2019.methods$Water.s.Xevo!="Failed",
  "Water.s.Xevo"]
  
# AB.Sciex.Qtrap:
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$AB.Sciex.Qtrap),
  "AB.Sciex.Qtrap"] <-  "Failed"
wambaugh2019.methods[wambaugh2019.methods$AB.Sciex.Qtrap!="Failed",
  "Instrument"] <- "AB Sciex QTRAP"
wambaugh2019.methods[wambaugh2019.methods$AB.Sciex.Qtrap!="Failed",
  "Analysis.Notes"] <- 
  wambaugh2019.methods[wambaugh2019.methods$AB.Sciex.Qtrap!="Failed",
  "AB.Sciex.Qtrap"]  

# Agilent.GCMS:
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$Agilent.GCMS),
  "Agilent.GCMS"] <-  "Failed"
wambaugh2019.methods[wambaugh2019.methods$Agilent.GCMS!="Failed",
  "Instrument"] <- "Agilent GCMS"
wambaugh2019.methods[wambaugh2019.methods$Agilent.GCMS!="Failed",
  "Analysis.Notes"] <- 
  wambaugh2019.methods[wambaugh2019.methods$Agilent.GCMS!="Failed",
  "Agilent.GCMS"]  

# GCTOF:
# Need to convert NA's to something to allow logic to work on whole column:
wambaugh2019.methods[is.na(wambaugh2019.methods$GCTOF),
  "GCTOF"] <-  "Failed"
wambaugh2019.methods[wambaugh2019.methods$GCTOF!="Failed",
  "Instrument"] <- "GCTOF"
wambaugh2019.methods[wambaugh2019.methods$GCTOF!="Failed",
  "Analysis.Notes"] <- 
  wambaugh2019.methods[wambaugh2019.methods$GCTOF!="Failed",
  "GCTOF"]  
  
# Add other needed columns:
wambaugh2019.methods$ISTD.Name <- "Bucetin and Diclofenac"

create_method_table(wambaugh2019.methods,
  compound.col="PREFERRED_NAME",
  analysis.method.col="Method",
  analysis.instrument.col="Instrument", 
  analysis.parameters.col="Analysis.Notes")

}
\author{
John Wambaugh
}
