% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_caco2.R
\name{format_caco2}
\alias{format_caco2}
\title{Creates a standardized data table reporting Caco-2 data}
\usage{
format_caco2(
  data.in,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  series.col = "Series",
  series = NULL,
  compound.col = "Compound.Name",
  area.col = "Area",
  istd.col = "ISTD.Area",
  type.col = "Type",
  direction.col = "Direction",
  membrane.area.col = "Membrane.Area",
  membrane.area = NULL,
  receiver.vol.col = "Vol.Receiver",
  donor.vol.col = "Vol.Donor",
  compound.conc = NULL,
  compound.conc.col = "Nominal.Conc",
  cal = NULL,
  cal.col = "Cal",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  meas.time.col = "Time",
  meas.time = 2,
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  nominal.test.conc = NULL,
  nominal.test.conc.col = "Test.Target.Conc",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters"
)
}
\arguments{
\item{data.in}{A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{sample.col}{Which column of data.in indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{lab.compound.col}{Which column of data.in indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{dtxsid.col}{Which column of data.in indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{date.col}{Which column of data.in indicates the laboratory measurment
date (Defaults to "Date")}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate with the same analytical chemistry 
(Defaults to "Series")}

\item{series}{If this argument is used (defaults to NULL) every observation 
in the table is assigned the value of the argument and the corresponding
column in input.table (if present) is ignored.}

\item{compound.col}{Which column of data.in indicates the test compound
(Defaults to "Compound.Name")}

\item{area.col}{Which column of data.in indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{istd.col}{Which column of data.in indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{type.col}{Which column of data.in indicates the direction of the 
measurements (either "AtoB" for apical to basolateral or "BtoA" for vice 
versa) (Defaults to "Direction")}

\item{direction.col}{Which column of data.in indicates the direction of
the Caco-2 permeability experiment, either apical to basal (AtoB) or basal
to aprical (BtoA). (Defaults to "Direction")}

\item{membrane.area.col}{Which column of data.in indicates the area of the
Caco-2 monolayer (in cm^2) (Defaults to "Membrane.Area")}

\item{membrane.area}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{receiver.vol.col}{Which column of data.in indicates the volume 
(in cm^3) of the receiver portion of the Caco-2 experimental well
(Defaults to "Vol.Receiver")}

\item{donor.vol.col}{Which column of data.in indicates the volume 
(in cm^3) of the donor portion of the Caco-2 experimental well where the
test chemical is added
(Defaults to "Vol.Donor")}

\item{cal}{If this argument is used (defaults to NULL) every observation in
the table is assigned the value of the argument and the corresponding
column in input.table (if present) is ignored.

#param compound.conc.col Which column indicates the intended concentration 
of the test chemical for calibration curves (Defaults to "Standard.Conc")}

\item{cal.col}{Which column of data.in indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{dilution}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{dilution.col}{Which column of data.in indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{meas.time.col}{Which column of data.in indicates the amount of time
before the receiver and donor compartments are measured (Defaults to "Time")}

\item{meas.time}{If this argument is used (defaults to 2 h) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.name}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.name.col}{Which column of data.in indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.conc.col}{Which column of data.in indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{nominal.test.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{nominal.test.conc.col}{Which column of data.in indicates the intended
test chemical concentration at time zero in the dosing solution (added to the
donor side of the Caco-2 test well) (Defaults to "Test.Target.Conc")}

\item{analysis.method}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.method.col}{Which column of data.in indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.instrument.col}{Which column of data.in indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.parameters.col}{Which column of data.in indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}
}
\value{
data.frame A data.frame in standardized "level1" format
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of membrane 
permeability using Caco-2 cells (Hubatsch, 2007). The input data frame is 
organized into a standard set of columns and is written to a tab-separated 
text file.
}
\details{
In this experiment an 
in vitro well is separated into two by a membrane composed of a monolayer of
Caco-2 cells. A test chemical is added to either the apical or basal side of
of the monolayer at time 0, and after a set time samples are taken from both
the "donor" (side where the test chemical was added) and the "receiver side.
Depending on the direction of the test the donor side can be either apical or
basal. 

The data frame of observations should be annotated according to direction 
(either apical to basal -- "AtoB" -- or basal to apical -- "BtoA") and type
of concentration measured:
\tabular{rr}{
  Blank with no chemical added \tab Blank \cr
  Dosing vehicle (C0) at target concentration \tab D0\cr
  Donor compartment at end of experiment \tab D2\cr
  Receiver compartment at end of experiment\tab R2\cr
}

Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{
library(invitroTKstats)
level0 <- TO1caco2
level1 <- format_caco2(level0,
  FILENAME="EPACyprotex2021",
  compound.col="CompoundName",
  compound.conc.col="Standard.Conc",
  membrane.area.col=0.11
  )

}
\references{
\insertRef{hubatsch2007determination}{invitroTKstats}
}
\author{
John Wambaugh
}
