% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_caco2.R
\name{format_caco2}
\alias{format_caco2}
\title{Creates a Standardized Data Table with Caco-2 Data (Level-1)}
\usage{
format_caco2(
  FILENAME = "MYDATA",
  data.in,
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  series = NULL,
  series.col = "Series",
  compound.col = "Compound.Name",
  area.col = "Area",
  istd.col = "ISTD.Area",
  type.col = "Type",
  direction.col = "Direction",
  membrane.area = NULL,
  membrane.area.col = "Membrane.Area",
  receiver.vol.col = "Vol.Receiver",
  donor.vol.col = "Vol.Donor",
  compound.conc = NULL,
  compound.conc.col = "Nominal.Conc",
  cal = NULL,
  cal.col = "Cal",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  meas.time = 2,
  meas.time.col = "Time",
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  nominal.test.conc = NULL,
  nominal.test.conc.col = "Test.Target.Conc",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters"
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the output Level-1 file.
"<FILENAME>-Caco-2-Level1.tsv". (Defaults to "MYDATA".)}

\item{data.in}{(Data Frame) A Level-0 data frame containing
mass-spectrometry peak areas, indication of chemical identity,
and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{sample.col}{(Character) Column name of \code{data.in} containing the unique mass
spectrometry (MS) sample name used by the laboratory. (Defaults to
"Lab.Sample.Name".)}

\item{lab.compound.col}{(Character) Column name of \code{data.in} containing the test compound
name used by the laboratory. (Defaults to "Lab.Compound.Name".)}

\item{dtxsid.col}{(Character) Column name of \code{data.in} containing EPA's DSSTox Structure
ID (\url{http://comptox.epa.gov/dashboard}). (Defaults to "DTXSID".)}

\item{date.col}{(Character) Column name of \code{data.in} containing the laboratory measurement
date. (Defaults to "Date".)}

\item{series}{(Numeric) Index of simultaneous replicates with the same analytical chemistry. 
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all tested compounds 
use the same number of replicates.)}

\item{series.col}{(Character) Column name containing \code{series} information.
(Defaults to "Series".) 
(Note: \code{data.in} does not necessarily have this field. 
If this field is missing, it can be auto-filled with the value 
specified in \code{series}.)}

\item{compound.col}{(Character) Column name of \code{data.in} containing the test compound.
(Defaults to "Compound.Name".)}

\item{area.col}{(Character) Column name of \code{data.in} containing the target analyte (that
is, the test compound) MS peak area. (Defaults to "Area".)}

\item{istd.col}{(Character) Column name of \code{data.in} containing the
MS peak area for the internal standard. (Defaults to "ISTD.Area".)}

\item{type.col}{(Character) Column name of \code{data.in} containing the sample type (see table
under Details). (Defaults to "Type".)}

\item{direction.col}{(Character) Column name of \code{data.in} containing the direction of
the Caco-2 permeability experiment: either apical donor to basal receiver (AtoB), or 
basal donor to apical receiver (BtoA). (Defaults to "Direction".)}

\item{membrane.area}{(Numeric) The area of the Caco-2 monolayer (in cm^2). 
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all tested compounds 
have the same area for the Caco-2 monolayer.)}

\item{membrane.area.col}{(Character) Column name containing \code{membrane.area} 
information. (Defaults to "Membrane.Area".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{membrane.area}.)}

\item{receiver.vol.col}{(Character) Column name of \code{data.in} containing the volume
(in cm^3) of the receiver portion of the Caco-2 experimental well. 
(Defaults to "Vol.Receiver".)}

\item{donor.vol.col}{(Character) Column name of \code{data.in} containing the volume
(in cm^3) of the donor portion of the Caco-2 experimental well where the
test chemical is added. (Defaults to "Vol.Donor".)}

\item{compound.conc}{(Numeric) The concentration
of the test chemical for calibration curves. (Note: Single entry only, 
use only if all test compounds have the same concentration for calibration curves.) 
(Defaults to \code{NULL}.)}

\item{compound.conc.col}{(Character) Column name containing \code{compound.conc} 
information. (Defaults to "Nominal.Conc".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{compound.conc}.)}

\item{cal}{(Character) MS calibration the samples were based on. Typically, this uses 
indices or dates to represent if the analyses were done on different machines on 
the same day or on different days with the same MS analyzer. (Defaults to \code{NULL}.) 
(Note: Single entry only, 
use only if all data were collected based on the same calibration.)}

\item{cal.col}{(Character) Column name containing \code{cal} 
information. (Defaults to "Cal".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{cal}.)}

\item{dilution}{(Numeric) Number of times the sample was diluted before MS 
analysis. (Defaults to \code{NULL}.) (Note: Single entry only, use only if all 
samples underwent the same number of dilutions.)}

\item{dilution.col}{(Character) Column name containing \code{dilution} 
information. (Defaults to "Dilution.Factor".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{dilution}.)}

\item{meas.time}{(Numeric) The amount of time (in hours) before the receiver and donor 
compartments are measured. (Defaults to 2.)}

\item{meas.time.col}{(Character) Column name containing \code{meas.time} 
information. (Defaults to "Time".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{meas.time}.)}

\item{istd.name}{(Character) The identity of the internal standard. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if all tested compounds use the same internal standard.)}

\item{istd.name.col}{(Character) Column name containing \code{istd.name} information. (Defaults to "ISTD.Name".) 
(Note: \code{data.in} does not necessarily have this field. If this field is missing, 
it can be auto-filled with the value specified in \code{istd.name}.)}

\item{istd.conc}{(Numeric) The concentration for the internal standard. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if all tested compounds have the same 
internal standard concentration.)}

\item{istd.conc.col}{(Character) Column name containing \code{istd.conc} information. (Defaults to "ISTD.Conc".) 
(Note: \code{data.in} does not necessarily have this field. If this field is missing, 
it can be auto-filled with the value specified in \code{istd.conc}.)}

\item{nominal.test.conc}{(Numeric) The test chemical concentration in the dosing solution
that is added to the donor side at time zero. (Defaults to \code{NULL}.)
(Note: Single entry only, use only if the same initial 
concentration was used for all tested compounds.)}

\item{nominal.test.conc.col}{(Character) Column name containing \code{nominal.test.conc} 
information. (Defaults to "Test.Target.Conc".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{nominal.test.conc}.)}

\item{analysis.method}{(Character) The analytical chemistry analysis method, 
typically "LCMS" or "GCMS", liquid chromatography or gas chromatography–mass
spectrometry, respectively. (Defaults to \code{NULL}.)
(Note: Single entry only, use only if the same method was used for all tested compounds.)}

\item{analysis.method.col}{(Character) Column name containing \code{analysis.method} 
information. (Defaults to "Analysis.Method".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{analysis.method}.)}

\item{analysis.instrument}{(Character) The instrument used for chemical analysis, 
for example "Agilent 6890 GC with model 5973 MS". (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if the same instrument was used for all tested compounds.)}

\item{analysis.instrument.col}{(Character) Column name containing \code{analysis.instrument} 
information. (Defaults to "Analysis.Instrument".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{analysis.instrument}.)}

\item{analysis.parameters}{(Character) The parameters used to identify the 
compound on the chemical analysis instrument, for example
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0". (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if the same parameters were used for all tested compounds.)}

\item{analysis.parameters.col}{(Character) Column name containing \code{analysis.parameters} 
information. (Defaults to "Analysis.Parameters".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{analysis.parameters}.)}
}
\value{
A Level-1 data frame with a standardized format containing a  
standardized set of columns and column names with membrane permeability data
from a Caco-2 assay.
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of membrane
permeability using Caco-2 cells \insertCite{hubatsch2007determination}{invitroTKstats}.
The input data frame is organized into a standard set of columns and is
written to a tab-separated text file.
}
\details{
In this experiment an
in vitro well is separated into two by a membrane composed of a monolayer of
Caco-2 cells. A test chemical is added to either the apical or basal side of
of the monolayer at time 0, and after a set time samples are taken from both
the "donor" (side where the test chemical was added) and the "receiver side.
Depending on the direction of the test the donor side can be either apical or
basal.

The data frame of observations should be annotated according to direction
(either apical to basal -- "AtoB" -- or basal to apical -- "BtoA") and type
of concentration measured:
\tabular{rr}{
  Blank with no chemical added \tab Blank \cr
  Dosing vehicle (C0) at target concentration \tab D0\cr
  Donor compartment at end of experiment \tab D2\cr
  Receiver compartment at end of experiment\tab R2\cr
}

Chemical concentration is calculated qualitatively as a response and 
returned as a column in the output data frame:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{
library(invitroTKstats)
level0 <- TO1caco2
level1 <- format_caco2(level0,
                       FILENAME="EPACyprotex2021",
                       sample.col="SampleName",
                       dtxsid.col="CompoundName",
                       lab.compound.col="CompoundName",
                       cal=1,
                       istd.conc.col="ISTD.Conc",
                       compound.col="CompoundName",
                       compound.conc.col="Test.Target.Conc",
                       membrane.area=0.11,
                       series=1,
                       analysis.parameters="Feature",
                       analysis.instrument="GC or LC",
                       analysis.method="Mass Spec"
                      )

}
\references{
\insertRef{hubatsch2007determination}{invitroTKstats}
}
\author{
John Wambaugh
}
