% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clint.R
\name{format_clint}
\alias{format_clint}
\title{Creates a standardized data table reporting hepatocye clearance data}
\usage{
format_clint(
  clint.data,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  date.col = "Date",
  compound.col = "Compound.Name",
  dtxsid.col = "DTXSID",
  lab.compound.col = "Lab.Compound.Name",
  type.col = "Sample.Type",
  dilution.col = "Dilution.Factor",
  cal.col = "Cal",
  istd.name.col = "ISTD.Name",
  istd.conc.col = "ISTD.Conc",
  istd.col = "ISTD.Area",
  density.col = "Hep.Density",
  conc.col = "Conc",
  time.col = "Time",
  area.col = "Area",
  analysis.method.col = "Analysis.Method",
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters.col = "Analysis.Parameters"
)
}
\arguments{
\item{clint.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{sample.col}{Which column of clint.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{date.col}{Which column of clint.data indicates the laboratory measurment
date (Defaults to "Date")}

\item{compound.col}{Which column of clint.data indicates the test compound
(Defaults to "Compound.Name")}

\item{dtxsid.col}{Which column of clint.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{lab.compound.col}{Which column of clint.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{type.col}{Which column of clint.data indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{dilution.col}{Which column of clint.data indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{cal.col}{Which column of clint.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{istd.name.col}{Which column of clint.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc.col}{Which column of clint.data indicates the concentration 
(units if uM) of
the internal standard (Defaults to "ISTD.Conc")}

\item{istd.col}{Which column of clint.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{density.col}{Which column of clint.data indicates the density (units of
millions of hepatocytes per mL) hepatocytes in the in vitro incubation 
(Defaults to "Hep.Density" )}

\item{conc.col}{Which column of clint.data indicates the intended
test chemical concentration 
(units if uM) of
at time zero (Defaults to "Conc")}

\item{time.col}{Which column of clint.data indicates the intended
time of the measurment (in minutes) since the test chemical was introduced
into the hepatocyte incubation (Defaults to "Time")}

\item{area.col}{Which column of clint.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{analysis.method.col}{Which column of PPB.data indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument.col}{Which column of PPB.data indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters.col}{Which column of PPB.data indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}

\item{series.col}{Which column of clint.data indicates the "series", that is
a simultaneous replicate (Defaults to "Series")}
}
\value{
\item{data.frame}{A data.frame in standardized "level1" format}
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical stability
when incubated with suspended hepatocytes (Shibata, 2000). Disappareance of 
the chemical over time is assumed to be due to metabolism by the hepatocytes.
An input dataframe is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank \tab Blank\cr
  Hepatocyte inciubation concentration \tab Cvst\cr
}
Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{

library(invitroTKstats)

clint <- wambaugh2019.clint
clint$Date <- "2019"
clint$Sample.Type <- "Blank"
clint$Time..mins. <- as.numeric(clint$Time..mins.)
clint[!is.na(clint$Time..mins.),"Sample.Type"] <- "Cvst"
clint$ISTD.Name <- "Bucetin, Propranolol, and Diclofenac"
clint$ISTD.Conc <- 1
clint$Dilution.Factor <- 1
clint[is.na(clint$FileName),"FileName"]<-"Wambaugh2019"
clint$Hep.Density <- 0.5

level1 <- format_clint(clint,
  FILENAME="Wambaugh2019",
  sample.col="Sample.Name",
  compound.col="Preferred.Name",
  lab.compound.col="Name",
  time.col="Time..mins.",
  cal.col="FileName")

}
\references{
Shibata, Yoshihiro, Hiroyuki Takahashi, and Yasuyuki Ishii. "A convenient in 
vitro screening method for predicting in vivo drug metabolic clearance using 
isolated hepatocytes suspended in serum." Drug metabolism and disposition 
28.12 (2000): 1518-1523.
}
\author{
John Wambaugh
}
