% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clint.R
\name{format_clint}
\alias{format_clint}
\title{Creates a Standardized Data Table Reporting Hepatocyte Clearance Data (Level-1)}
\usage{
format_clint(
  data.in,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  date.col = "Date",
  compound.col = "Compound.Name",
  dtxsid.col = "DTXSID",
  lab.compound.col = "Lab.Compound.Name",
  type.col = "Sample.Type",
  density = NULL,
  density.col = "Hep.Density",
  cal = NULL,
  cal.col = "Cal",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  time.col = "Time",
  time = 2,
  istd.col = "ISTD.Area",
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  std.conc = NULL,
  std.conc.col = "Standard.Conc",
  clint.assay.conc = NULL,
  clint.assay.conc.col = "Clint.Assay.Conc",
  area.col = "Area",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters",
  note.col = "Note",
  level0.file.col = "Level0.File",
  level0.file = NULL,
  level0.sheet.col = "Level0.Sheet",
  level0.sheet = NULL
)
}
\arguments{
\item{data.in}{(Data Frame) A data frame or a matrix containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{(Character) A string used to identify the output Level-1 file.
"<FILENAME>-Clint-Level1.tsv". (Defaults to "MYDATA").}

\item{sample.col}{(Character) Column name of \code{data.in} containing the unique mass
spectrometry (MS) sample name used by the laboratory. (Defaults to
"Lab.Sample.Name".)}

\item{date.col}{(Character) Column name of \code{data.in} containing the laboratory measurement
date. (Defaults to "Date".)}

\item{compound.col}{(Character) Column name of \code{data.in} containing the test compound.
(Defaults to "Compound.Name".)}

\item{dtxsid.col}{(Character) Column name of \code{data.in} containing EPA's DSSTox Structure
ID (\url{http://comptox.epa.gov/dashboard}). (Defaults to "DTXSID".)}

\item{lab.compound.col}{(Character) Column name of \code{data.in} containing the test compound
name used by the laboratory. (Defaults to "Lab.Compound.Name".)}

\item{type.col}{(Character) Column name of \code{data.in} containing the sample type (see table
under Details). (Defaults to "Sample.Type".)}

\item{density}{(Numeric) The density (units of
millions of hepatocytes per mL) hepatocytes in the in vitro incubation. 
(Defaults to \code{NULL}.) (Note: Single entry only, 
use only if all tested compounds have the same density.)}

\item{density.col}{(Character) Column name containing \code{density} 
information. (Defaults to "Hep.Density".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{density}.)}

\item{cal}{(Character) MS calibration the samples were based on. Typically, this uses 
indices or dates to represent if the analyses were done on different machines on 
the same day or on different days with the same MS analyzer. (Defaults to \code{NULL}.) 
(Note: Single entry only, 
use only if all data were collected based on the same calibration.)}

\item{cal.col}{(Character) Column name containing \code{cal} 
information. (Defaults to "Cal".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{cal}.)}

\item{dilution}{(Numeric) Number of times the sample was diluted before MS 
analysis. (Defaults to \code{NULL}.) (Note: Single entry only, use only if all 
samples underwent the same number of dilutions.)}

\item{dilution.col}{(Character) Column name containing \code{dilution} 
information. (Defaults to "Dilution.Factor".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{dilution}.)}

\item{time.col}{(Character) Column name containing \code{time} 
information. (Defaults to "Time".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{time}.)}

\item{time}{(Numeric) The intended time of the measurement (in minutes) since the test 
chemicals was introduced into the hepatocyte incubation. (Defaults to 2.)}

\item{istd.col}{(Character) Column name of data.in containing the
MS peak area for the internal standard. (Defaults to "ISTD.Area".)}

\item{istd.name}{(Character) The identity of the internal standard. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if all tested compounds use the same internal standard.)}

\item{istd.name.col}{(Character) Column name containing \code{istd.name} information. 
(Defaults to "ISTD.Name".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{istd.name}.)}

\item{istd.conc}{(Numeric) The concentration for the internal standard. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if all tested compounds have the same 
internal standard concentration.)}

\item{istd.conc.col}{(Character) Column name containing \code{istd.conc} information. 
(Defaults to "ISTD.Conc".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{istd.conc}.)}

\item{std.conc}{(Numeric) The standard test chemical concentration for 
the intrinsic clearance assay. (Defaults to \code{NULL}.) (Note: Single entry only, 
use only if the same standard concentration was used for all tested compounds.)}

\item{std.conc.col}{(Character) Column name containing \code{std.conc} 
information. (Defaults to "Standard.Conc".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{std.conc}.)}

\item{clint.assay.conc}{(Numeric) The initial test chemical concentration for 
the intrinsic clearance assay. (Defaults to \code{NULL}.) (Note: Single entry only, 
use only if the same initial concentration was used for all tested compounds.)}

\item{clint.assay.conc.col}{(Character) Column name containing \code{clint.assay.conc} 
information. (Defaults to "Clint.Assay.Conc".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{clint.assay.conc}.)}

\item{area.col}{(Character) Column name of data.in containing the target analyte (that
is, the test compound) MS peak area. (Defaults to "Area".)}

\item{analysis.method}{(Character) The analytical chemistry analysis method, 
typically "LCMS" or "GCMS", liquid chromatography or gas chromatography–mass spectrometry, respectively. 
(Defaults to \code{NULL}.) (Note: Single entry only, 
use only if the same method was used for all tested compounds.)}

\item{analysis.method.col}{(Character) Column name containing \code{analysis.method} 
information. (Defaults to "Analysis.Method".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{analysis.method}.)}

\item{analysis.instrument}{(Character) The instrument used for chemical analysis, 
for example "Waters Xevo TQ-S micro (QEB0036)". (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if the same instrument was used for all tested compounds.)}

\item{analysis.instrument.col}{(Character) Column name containing \code{analysis.instrument} 
information. (Defaults to "Analysis.Instrument".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{analysis.instrument}.)}

\item{analysis.parameters}{(Numeric) The parameters used to identify the 
compound on the chemical analysis instrument. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if the same parameters were used for all tested compounds.)}

\item{analysis.parameters.col}{(Character) Column name containing \code{analysis.parameters} 
information. (Defaults to "Analysis.Parameters".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{analysis.parameters}.)}

\item{note.col}{(Character) Column name of data.in containing additional notes on 
test compounds. (Defaults to "Note").}

\item{level0.file.col}{(Character) Column name containing \code{level0.file} information. 
(Defaults to "Level0.File".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{level0.file}.)}

\item{level0.file}{(Character) The Level-0 file from which the data.in were obtained.
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all rows in data.in
were obtained from the same Level-0 file.)}

\item{level0.sheet.col}{(Character) Column name containing \code{level0.sheet} information.
(Defaults to "Level0.Sheet".) (Note: \code{data.in} does not
necessarily have this field. If this field is missing, it can be auto-filled with the value 
specified in \code{level0.sheet}.)}

\item{level0.sheet}{(Character) The specific sheet name of Level-0 file from which the 
data.in is obtained from, if the level-0 file is an Excel workbook. 
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all rows in data.in
were obtained from the same sheet in the same Level-0 file.)}
}
\value{
A data frame in standardized Level-1 format containing a  
standardized set of columns with standardized column names.
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical stability
when incubated with suspended hepatocytes \insertCite{shibata2002prediction}{invitroTKstats}. Disappearance of
the chemical over time is assumed to be due to metabolism by the hepatocytes.
The input data frame is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank \tab Blank\cr
  Hepatocyte incubation concentration \tab Cvst\cr
  Inactivated Hepatocytes \tab Inactive\cr
  Calibration Curve \tab CC\cr
}

Chemical concentration is calculated qualitatively as a response and 
returned as a column in the output data frame:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{

library(invitroTKstats)

clint <- wambaugh2019.clint
clint$Date <- "2019"
clint$Sample.Type <- "Blank"
clint$Time..mins. <- as.numeric(clint$Time..mins.)
clint[!is.na(clint$Time..mins.),"Sample.Type"] <- "Cvst"
clint$ISTD.Name <- "Bucetin, Propranolol, and Diclofenac"
clint$ISTD.Conc <- 1
clint$Dilution.Factor <- 1
clint[is.na(clint$FileName),"FileName"]<-"Wambaugh2019"
clint$Hep.Density <- 0.5

level1 <- format_clint(clint,
  FILENAME="Wambaugh2019",
  sample.col="Sample.Name",
  compound.col="Preferred.Name",
  lab.compound.col="Name",
  time.col="Time..mins.",
  cal.col="FileName")

}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}
}
\author{
John Wambaugh
}
