% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clint.R
\name{format_clint}
\alias{format_clint}
\title{Creates a standardized data table reporting hepatocye clearance data}
\usage{
format_clint(
  data.in,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  date = NULL,
  date.col = "Date",
  compound.col = "Compound.Name",
  dtxsid.col = "DTXSID",
  lab.compound.col = "Lab.Compound.Name",
  type.col = "Sample.Type",
  density = NULL,
  density.col = "Hep.Density",
  compound.conc = NULL,
  compound.conc.col = "Nominal.Conc",
  cal = NULL,
  cal.col = "Cal",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  time.col = "Time",
  time = 2,
  istd.col = "ISTD.Area",
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  std.conc = NULL,
  std.conc.col = "Standard.Conc",
  clint.assay.conc = NULL,
  clint.assay.conc.col = "Clint.Assay.Conc",
  area.col = "Area",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters",
  note.col = "Note",
  level0.file.col = "Level0.File",
  level0.file = NULL,
  level0.sheet.col = "Level0.Sheet",
  level0.sheet = NULL
)
}
\arguments{
\item{data.in}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{sample.col}{Which column of input.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{date.col}{Which column of input.data indicates the laboratory measurment
date (Defaults to "Date")}

\item{compound.col}{Which column of input.data indicates the test compound
(Defaults to "Compound.Name")}

\item{dtxsid.col}{Which column of input.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{lab.compound.col}{Which column of input.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{type.col}{Which column of input.data indicates the direction of the 
measurements (either "AtoB" for apical to basolateral or "BtoA" for vice 
versa) (Defaults to "Direction")}

\item{density.col}{A single value to be used for all samples indicating
the density of hepatocytes in suspension (10^6 hepatocytes / mL) 
(Defaults to NULL)}

\item{cal}{If this argument is used (defaults to NULL) every observation in
the table is assigned the value of the argument and the corresponding
column in input.table (if present) is ignored.

#param compound.conc.col Which column indictes the intended concentration 
of the test chemical for calibration curves (Defaults to "Standard.Conc")}

\item{cal.col}{Which column of input.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{dilution}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{dilution.col}{Which column of data.in indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{time.col}{Which column of data.in indicates the intended
time of the measurment (in minutes) since the test chemical was introduced
into the hepatocyte incubation (Defaults to "Time")}

\item{istd.col}{Which column of input.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{istd.name}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.name.col}{Which column of input.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.conc.col}{Which column of input.data indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{area.col}{Which column of input.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{analysis.method}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.method.col}{Which column of input.data indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.instrument.col}{Which column of input.data indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.parameters.col}{Which column of input.data indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}

\item{level0.file.col}{Which column of PPB.data indicates the file from
which the data were obtained (for example "MyWorkbook.xlsx").}

\item{level0.file}{If this argument is used (defaults to NULL) every
observation in the table is assigned the value of the argument and the
corresponding column in input.table (if present) is ignored.}

\item{level0.sheet.col}{Which column of PPB.data indicates the specific 
sheet containing the data if the file is an Excel workbook}

\item{level0.sheet}{If this argument is used (defaults to NULL) every
observation in the table is assigned the value of the argument and the
corresponding column in input.table (if present) is ignored.}

\item{conc.col}{Which column of data.in indicates the intended
test chemical concentration 
(units if uM) of
at time zero (Defaults to "Conc")}

\item{input.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type. The data frame should
contain columns with names specified by the following arguments:}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate with the same analytical chemistry 
(Defaults to "Series")}

\item{series}{If this argument is used (defaults to NULL) every observation 
in the table is assigned the value of the argument and the corresponding
column in input.table (if present) is ignored.}

\item{nominal.test.conc.col}{Which column of input.data indicates the intended
test chemical concentration at time zero in the dosing solution (added to the
donor side of the Caco-2 test well) (Defaults to "Test.Target.Conc")}

\item{nominal.test.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}
}
\value{
\item{data.frame}{A data.frame in standardized "level1" format}
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical stability
when incubated with suspended hepatocytes (Shibata, 2000). Disappareance of 
the chemical over time is assumed to be due to metabolism by the hepatocytes.
An input dataframe is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank \tab Blank\cr
  Hepatocyte inciubation concentration \tab Cvst\cr
  Inactivated Hepatocytes \tab Inactive\cr
  Calibration Curve \tab CC\cr
}
Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{

library(invitroTKstats)

clint <- wambaugh2019.clint
clint$Date <- "2019"
clint$Sample.Type <- "Blank"
clint$Time..mins. <- as.numeric(clint$Time..mins.)
clint[!is.na(clint$Time..mins.),"Sample.Type"] <- "Cvst"
clint$ISTD.Name <- "Bucetin, Propranolol, and Diclofenac"
clint$ISTD.Conc <- 1
clint$Dilution.Factor <- 1
clint[is.na(clint$FileName),"FileName"]<-"Wambaugh2019"
clint$Hep.Density <- 0.5

level1 <- format_clint(clint,
  FILENAME="Wambaugh2019",
  sample.col="Sample.Name",
  compound.col="Preferred.Name",
  lab.compound.col="Name",
  time.col="Time..mins.",
  cal.col="FileName")

}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}
}
\author{
John Wambaugh
}
