% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fup_red.R
\name{format_fup_red}
\alias{format_fup_red}
\title{Creates a Standardized Data Frame Reporting Rapid Equilibrium Dialysis (RED) PPB data}
\usage{
format_fup_red(
  data.in,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  date.col = "Date",
  compound.col = "Compound.Name",
  dtxsid.col = "DTXSID",
  lab.compound.col = "Lab.Compound.Name",
  type.col = "Sample.Type",
  cal = NULL,
  cal.col = "Cal",
  replicate = NULL,
  replicate.col = "Replicate",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  time.col = "Time",
  istd.col = "ISTD.Area",
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  test.nominal.conc = NULL,
  test.nominal.conc.col = "Test.Target.Conc",
  plasma.percent = NULL,
  plasma.percent.col = "Plasma.Percent",
  std.conc = NULL,
  std.conc.col = "Standard.Conc",
  area.col = "Area",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters",
  note.col = "Note",
  level0.file.col = "Level0.File",
  level0.file = NULL,
  level0.sheet.col = "Level0.Sheet",
  level0.sheet = NULL
)
}
\arguments{
\item{data.in}{(Data Frame) A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{(Character) A string used to identify the output Level-1 file.
"<FILENAME>-fup-RED-Level1.tsv". (Defaults to "MYDATA".)}

\item{sample.col}{(Character) Column name of data.in containing the unique mass
spectrometry (MS) sample name used by the laboratory. (Defaults to
"Lab.Sample.Name".)}

\item{date.col}{(Character) Column name of data.in containing the laboratory measurement
date. (Defaults to "Date".)}

\item{compound.col}{(Character) Column name of data.in containing the test compound.
(Defaults to "Compound.Name".)}

\item{dtxsid.col}{(Character) Column name of data.in containing EPA's DSSTox Structure
ID (\url{http://comptox.epa.gov/dashboard}). (Defaults to "DTXSID".)}

\item{lab.compound.col}{(Character) Column name of data.in containing the test compound
name used by the laboratory. (Defaults to "Lab.Compound.Name".)}

\item{type.col}{(Character) Column name of data.in containing the sample type (see table
under Details). (Defaults to "Sample.Type".)}

\item{cal}{(Character) MS calibration the samples were based on, typically uses 
indices or dates to represent if the analyses were done on different machines on 
the same day or on different days with the same MS analyzer. (Defaults to \code{NULL}.) 
(Note: Single entry only, 
use only if all data were collected based on the same calibration.)}

\item{cal.col}{(Character) Column name containing \code{cal} 
information. (Defaults to "Cal".)}

\item{replicate}{(Numeric) Simultaneous replicates with the same analytical chemistry. 
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all tested compounds 
use the same number of replicates.)}

\item{replicate.col}{(Character) Column name containing \code{replicate} 
information. (Defaults to "Replicate".)}

\item{dilution}{(Numeric) Number of times the sample was diluted before MS 
analysis. (Defaults to \code{NULL}.) (Note: Single entry only, use only if all 
samples underwent the same number of dilutions.)}

\item{dilution.col}{(Character) Column name containing \code{dilution} 
information. (Defaults to "Dilution.Factor".)}

\item{time.col}{(Character) Column name of data.in containing the time. 
(Defaults to "Time".)}

\item{istd.col}{(Character) Column name of data.in containing the
MS peak area for the internal standard. (Defaults to "ISTD.Area".)}

\item{istd.name}{(Character) The identity of the internal standard. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if all tested compounds use the same internal standard.)}

\item{istd.name.col}{(Character) Column name containing \code{istd.name} information. 
(Defaults to "ISTD.Name".)}

\item{istd.conc}{(Numeric) The concentration for the internal standard. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if all tested compounds have the same 
internal standard concentration.)}

\item{istd.conc.col}{(Character) Column name containing \code{istd.conc} information. 
(Defaults to "ISTD.Conc".)}

\item{test.nominal.conc}{(Numeric) The intended test chemical concentration 
at time zero. (Defaults to \code{NULL}.) (Note: Single entry only, use only 
if all tested compounds used the same concentration at time zero.)}

\item{plasma.percent}{(Numeric) Plasma Percent. (Defaults to \code{NULL}.) (Note: Single entry only, use only 
if all tested compounds have the same plasma percent.)}

\item{plasma.percent.col}{(Character) Column name containing \code{plasma.percent} 
information. (Defaults to "Plasma.Percent".)}

\item{std.conc}{(Numeric) The standard test chemical concentration for 
the intrinsic clearance assay. (Defaults to \code{NULL}.) (Note: Single entry only, 
use only if the same standard concentration was used for all tested compounds.)}

\item{std.conc.col}{(Character) Column name containing \code{std.conc} 
information. (Defaults to "Standard.Conc".)}

\item{area.col}{(Character) Column name of data.in containing the target analyte (that
is, the test compound) MS peak area. (Defaults to "Area".)}

\item{analysis.method}{(Character) The analytical chemistry analysis method, 
typically "LCMS" or "GCMS", liquid chromatography or gas chromatography–mass spectrometry, respectively. 
(Defaults to \code{NULL}.) (Note: Single entry only, 
use only if the same method was used for all tested compounds.)}

\item{analysis.method.col}{(Character) Column name containing \code{analysis.method} 
information. (Defaults to "Analysis.Method".)}

\item{analysis.instrument}{(Character) The instrument used for chemical analysis, 
for example "Waters ACQUITY I-Class UHPLC - Xevo TQ-S uTQMS". (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if the same instrument was used for all tested compounds.)}

\item{analysis.instrument.col}{(Character) Column name containing \code{analysis.instrument} 
information. (Defaults to "Analysis.Instrument".)}

\item{analysis.parameters}{(Character) The parameters used to identify the 
compound on the chemical analysis instrument. (Defaults to \code{NULL}.) 
(Note: Single entry only, use only if the same parameters were used for all tested compounds.)}

\item{analysis.parameters.col}{(Character) Column name containing \code{analysis.instrument} 
information. (Defaults to "Analysis.Instrument".)}

\item{note.col}{(Character) Column name of data.in containing additional notes on 
test compounds. (Defaults to "Note".)}

\item{level0.file.col}{(Character) Column name containing \code{level0.file} information. 
(Defaults to "Level0.File".)}

\item{level0.file}{(Character) The Level-0 file from which the data.in were obtained.
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all rows in data.in
were obtained from the same Level-0 file.)}

\item{level0.sheet.col}{(Character) Column name containing \code{level0.sheet} information.
(Defaults to "Level0.Sheet".)}

\item{level0.sheet}{(Character) The specific sheet name of Level-0 file from which the 
data.in is obtained from, if the level-0 file is an Excel workbook. 
(Defaults to \code{NULL}.) (Note: Single entry only, use only if all rows in data.in
were obtained from the same sheet in the same Level-0 file.)}

\item{Test.Nominal.Conc.col}{(Character) Column name containing \code{test.nominal.conc} 
information. (Defaults to "Test.Target.Conc".)}
}
\value{
A data frame in standardized Level-1 format containing a  
standardized set of columns with standardized column names.
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using rapid equilibrium dialysis
\insertCite{waters2008validation}{invitroTKstats}.
The input data frame is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
  Plasma stability sample \tab Stability\cr
  Equilibrium Control Well 1 \tab EC1\cr
  Equilibrium Control Well 2 \tab EC2\cr
  Calibration Curve \tab CC\cr
}
Chemical concentration is calculated qualitatively as a response and 
returned as a column in the output data frame:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{
library(invitroTKstats)
red <- wambaugh2019.red
red$Date <- "2019"
red$Sample.Type <- "Blank"
red <- subset(red,!is.na(SampleName))
red[regexpr("PBS",red$SampleName)!=-1,"Sample.Type"] <- "PBS"
red[regexpr("Plasma",red$SampleName)!=-1,"Sample.Type"] <- "Plasma"
red$Dilution.Factor <- NA
red$Dilution.Factor <- as.numeric(red$Dilution.Factor)
red[red$Sample.Type=="PBS","Dilution.Factor"] <- 2
red[red$Sample.Type=="Plasma","Dilution.Factor"] <- 5
red[regexpr("T0",red$SampleName)!=-1,"Sample.Type"] <- "T0"

red$Test.Target.Conc <- 5
red$ISTD.Name <- "Bucetin and Diclofenac"
red$ISTD.Conc <- 1
red$Series <- 1

level1 <- format_fup_red(red,
 FILENAME="Wambaugh2019",
 sample.col="SampleName",
 compound.col="Preferred.Name",
 lab.compound.col="CompoundName",
 cal.col="RawDataSet")

}
\references{
\insertRef{waters2008validation}{invitroTKstats}
}
\author{
John Wambaugh
}
