% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fup_red.R
\name{format_fup_red}
\alias{format_fup_red}
\title{Creates a standardized data table reporting RED PPB data}
\usage{
format_fup_red(
  PPB.data,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  compound.col = "Compound.Name",
  area.col = "Area",
  series.col = "Series",
  type.col = "Sample.Type",
  cal.col = "Cal",
  dilution.col = "Dilution.Factor",
  istd.col = "ISTD.Area",
  istd.name.col = "ISTD.Name",
  istd.conc.col = "ISTD.Conc",
  nominal.test.conc.col = "Test.Target.Conc"
)
}
\arguments{
\item{PPB.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{sample.col}{Which column of PPB.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{lab.compound.col}{Which column of PPB.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{dtxsid.col}{Which column of PPB.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{date.col}{Which column of PPB.data indicates the laboratory measurment
date (Defaults to "Date")}

\item{compound.col}{Which column of PPB.data indicates the test compound
(Defaults to "Compound.Name")}

\item{area.col}{Which column of PPB.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate (Defaults to "Series")}

\item{type.col}{Which column of PPB.data indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{cal.col}{Which column of PPB.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{dilution.col}{Which column of PPB.data indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{istd.col}{Which column of PPB.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{istd.name.col}{Which column of PPB.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc.col}{Which column of PPB.data indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{nominal.test.conc.col}{Which column of PPB.data indicates the intended
test chemical concentration at time zero (Defaults to "Test.Target.Conc")}
}
\value{
\item{data.frame}{A data.frame in standardized format}
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using rapid equilibrium dialysis (Waters, et al, 2008).
An input dataframe is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}
Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{
red <- wambaugh2019.red
red$Date <- "2019"
red$Sample.Type <- "Blank"
red <- subset(red,!is.na(SampleName))
red[regexpr("PBS",red$SampleName)!=-1,"Sample.Type"] <- "PBS"
red[regexpr("Plasma",red$SampleName)!=-1,"Sample.Type"] <- "Plasma"
red$Dilution.Factor <- NA
red$Dilution.Factor <- as.numeric(red$Dilution.Factor)
red[red$Sample.Type=="PBS","Dilution.Factor"] <- 2
red[red$Sample.Type=="Plasma","Dilution.Factor"] <- 5
red[regexpr("T0",red$SampleName)!=-1,"Sample.Type"] <- "T0"

red$Test.Target.Conc <- 5
red$ISTD.Name <- "Bucetin and Diclofenac"
red$ISTD.Conc <- 1
red$Series <- 1

level1 <- format_fup_red(red,
 FILENAME="Wambaugh2019",
 sample.col="SampleName",
 compound.col="Preferred.Name",
 lab.compound.col="CompoundName",
 cal.col="RawDataSet")

}
\references{
Waters, Nigel J., et al. "Validation of a rapid equilibrium dialysis 
approach for the measurement of plasma protein binding." Journal of 
Pharmaceutical Sciences 97.10 (2008): 4586-4595.
}
\author{
John Wambaugh
}
