% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fup_red.R
\name{format_fup_red}
\alias{format_fup_red}
\title{Creates a standardized data table reporting RED PPB data}
\usage{
format_fup_red(
  data.in,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  date = NULL,
  date.col = "Date",
  compound.col = "Compound.Name",
  dtxsid.col = "DTXSID",
  lab.compound.col = "Lab.Compound.Name",
  type.col = "Sample.Type",
  cal = NULL,
  cal.col = "Cal",
  replicate = NULL,
  replicate.col = "Replicate",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  time.col = "Time",
  time = 4,
  istd.col = "ISTD.Area",
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  test.nominal.conc = NULL,
  test.nominal.conc.col = "Test.Target.Conc",
  plasma.percent = NULL,
  plasma.percent.col = "Plasma.Percent",
  std.conc = NULL,
  std.conc.col = "Standard.Conc",
  area.col = "Area",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters",
  note.col = "Note",
  level0.file.col = "Level0.File",
  level0.file = NULL,
  level0.sheet.col = "Level0.Sheet",
  level0.sheet = NULL
)
}
\arguments{
\item{data.in}{A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{sample.col}{Which column of data.in indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{date.col}{Which column of data.in indicates the laboratory measurement
date (Defaults to "Date")}

\item{compound.col}{Which column of data.in indicates the test compound
(Defaults to "Compound.Name")}

\item{dtxsid.col}{Which column of data.in indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{lab.compound.col}{Which column of data.in indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{type.col}{Which column of data.in indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{cal.col}{Which column of data.in indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{dilution.col}{Which column of data.in indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{istd.col}{Which column of data.in indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{istd.name.col}{Which column of data.in indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc.col}{Which column of data.in indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{area.col}{Which column of data.in indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{analysis.method.col}{Which column of data.in indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument.col}{Which column of data.in indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters.col}{Which column of data.in indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}

\item{series.col}{Which column of data.in indicates the "series", that is
a simultaneous replicate (Defaults to "Series")}

\item{Test.Nominal.Conc.col}{Which column of data.in indicates the intended
test chemical concentration at time zero (Defaults to "Test.Target.Conc")}
}
\value{
\item{data.frame}{A data.frame in standardized "level1" format}
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using rapid equilibrium dialysis (Waters, et al, 2008).
An input data frame is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
  Plasma stability sample \tab Stability\cr
  Equilibrium Control Well 1 \tab EC1\cr
  Equilibrium Control Well 2 \tab EC2\cr
  Calibration Curve \tab CC\cr
}
Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{
library(invitroTKstats)
red <- wambaugh2019.red
red$Date <- "2019"
red$Sample.Type <- "Blank"
red <- subset(red,!is.na(SampleName))
red[regexpr("PBS",red$SampleName)!=-1,"Sample.Type"] <- "PBS"
red[regexpr("Plasma",red$SampleName)!=-1,"Sample.Type"] <- "Plasma"
red$Dilution.Factor <- NA
red$Dilution.Factor <- as.numeric(red$Dilution.Factor)
red[red$Sample.Type=="PBS","Dilution.Factor"] <- 2
red[red$Sample.Type=="Plasma","Dilution.Factor"] <- 5
red[regexpr("T0",red$SampleName)!=-1,"Sample.Type"] <- "T0"

red$Test.Target.Conc <- 5
red$ISTD.Name <- "Bucetin and Diclofenac"
red$ISTD.Conc <- 1
red$Series <- 1

level1 <- format_fup_red(red,
 FILENAME="Wambaugh2019",
 sample.col="SampleName",
 compound.col="Preferred.Name",
 lab.compound.col="CompoundName",
 cal.col="RawDataSet")

}
\references{
Waters, Nigel J., et al. "Validation of a rapid equilibrium dialysis 
approach for the measurement of plasma protein binding." Journal of 
Pharmaceutical Sciences 97.10 (2008): 4586-4595.
}
\author{
John Wambaugh
}
