% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fup_uc.R
\name{format_fup_uc}
\alias{format_fup_uc}
\title{Creates a standardized data table reporting UC PPB data}
\usage{
format_fup_uc(
  PPB.data,
  FILENAME = "MYDATA",
  sample.col = "Lab.Sample.Name",
  lab.compound.col = "Lab.Compound.Name",
  dtxsid.col = "DTXSID",
  date.col = "Date",
  compound.col = "Compound.Name",
  area.col = "Area",
  series.col = "Series",
  type.col = "Sample.Type",
  std.conc = NULL,
  std.conc.col = "Standard.Conc",
  cal = NULL,
  cal.col = "Cal",
  dilution = NULL,
  dilution.col = "Dilution.Factor",
  istd.col = "ISTD.Area",
  istd.name = NULL,
  istd.name.col = "ISTD.Name",
  istd.conc = NULL,
  istd.conc.col = "ISTD.Conc",
  uc.assay.conc = NULL,
  uc.assay.conc.col = "UC.Assay.Conc",
  analysis.method = NULL,
  analysis.method.col = "Analysis.Method",
  analysis.instrument = NULL,
  analysis.instrument.col = "Analysis.Instrument",
  analysis.parameters = NULL,
  analysis.parameters.col = "Analysis.Parameters",
  note.col = "Note",
  level0.file.col = "Level0.File",
  level0.file = NULL,
  level0.sheet.col = "Level0.Sheet",
  level0.sheet = NULL
)
}
\arguments{
\item{PPB.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identiy, and measurment type. The data frame should
contain columns with names specified by the following arguments:}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{sample.col}{Which column of PPB.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{lab.compound.col}{Which column of PPB.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{dtxsid.col}{Which column of PPB.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{date.col}{Which column of PPB.data indicates the laboratory measurment
date (Defaults to "Date")}

\item{compound.col}{Which column of PPB.data indicates the test compound
(Defaults to "Compound.Name")}

\item{area.col}{Which column of PPB.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate with the same analytical chemistry 
(Defaults to "Series")}

\item{type.col}{Which column of PPB.data indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{std.conc.col}{Which column indictes the intended concentration 
of the test chemical for calibration curves in uM (Defaults to "Standard.Conc")}

\item{cal}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{cal.col}{Which column of PPB.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{dilution}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{dilution.col}{Which column of PPB.data indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{istd.col}{Which column of PPB.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{istd.name}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.name.col}{Which column of PPB.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.conc.col}{Which column of PPB.data indicates the concentration of
the internal standard in uM (Defaults to "ISTD.Conc")}

\item{uc.assay.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{uc.assay.conc.col}{Which column indicates the intended initial
test chemical concentration in the UC assay in uM (Defaults to "Test.Target.Conc")}

\item{analysis.method}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.method.col}{Which column of PPB.data indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.instrument}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.instrument.col}{Which column of PPB.data indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.parameters}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.parameters.col}{Which column of PPB.data indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaulys to "Analysis.Paramaters").}

\item{level0.file.col}{Which column of PPB.data indicates the file from
which the data were obtained (for example "MyWorkbook.xlsx").}

\item{level0.file}{If this argument is used (defaults to NULL) every
observation in the table is assigned the value of the argument and the
corresponding column in input.table (if present) is ignored.}

\item{level0.sheet.col}{Which column of PPB.data indicates the specific 
sheet containing the data if the file is an Excel workbook}

\item{level0.sheet}{If this argument is used (defaults to NULL) every
observation in the table is assigned the value of the argument and the
corresponding column in input.table (if present) is ignored.}
}
\value{
data.frame A data.frame in standardized "level1" format
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using ultracentrifugation (Redgrave 1975?).
An input dataframe is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultra-centrifugation Aqueous Fraction \tab UC\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}
Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\examples{
library(invitroTKstats)
level0 <- kreutz2020
level1 <- format_fup_uc(level0,
  FILENAME="Kreutz2020",
  compound.col="Name",
  std.conc.col="Standard.Conc",
  area.col="Chem.Area"
  )

}
\references{
Redgrave, T. G., D. C. K. Roberts, and C. E. West. "Separation of plasma 
lipoproteins by density-gradient ultracentrifugation." Analytical 
Biochemistry 65.1-2 (1975): 42-49.#'
}
\author{
John Wambaugh
}
