% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Heaviside}
\alias{Heaviside}
\title{Heaviside}
\usage{
Heaviside(x, threshold = 0)
}
\arguments{
\item{x}{(Numeric) A numeric vector.}

\item{threshold}{(Numeric) A threshold value used to compare to elements in \code{x}. (Defaults to 0.)}
}
\value{
A vector of 1 and 0 in which 1 indicates the element in \code{x} is larger or equal to the threshold.
}
\description{
Heaviside
}
