% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Heaviside}
\alias{Heaviside}
\title{Heaviside}
\usage{
Heaviside(x, threshold = 0)
}
\arguments{
\item{x}{a vector}

\item{threshold}{defaults to 0}
}
\value{
a vector of 1 and 0, 1 indicates input element is above threshold
}
\description{
Heaviside
}
