% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initfunction_fup_uc}
\alias{initfunction_fup_uc}
\title{Set Initial Values for Fup UC Bayesian Model}
\usage{
initfunction_fup_uc(mydata, chain)
}
\arguments{
\item{mydata}{(List) Output of \code{build_mydata_clint}.}

\item{chain}{(Numeric) The number of Markov Chains to use.}
}
\value{
A list of initial values.
}
\description{
Set Initial Values for Fup UC Bayesian Model
}
