% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_level0.R
\name{merge_level0}
\alias{merge_level0}
\title{Merge multiple level0 files into a single table for processing}
\usage{
merge_level0(
  data.label = "MYDATA",
  level0.catalog,
  file.col = "File",
  sheet = NULL,
  sheet.col = "Sheet",
  skip.rows = NULL,
  skip.rows.col = "Skip.Rows",
  num.rows = NULL,
  num.rows.col = NULL,
  date = NULL,
  date.col = "Date",
  compound.col = "Chemical.ID",
  istd.col = "ISTD",
  sample.colname = NULL,
  sample.colname.col = "Sample.ColName",
  type.colname = NULL,
  type.colname.col = "Type",
  peak.colname = NULL,
  peak.colname.col = "Peak.ColName",
  istd.peak.colname = NULL,
  istd.peak.colname.col = "ISTD.Peak.ColName",
  conc.colname = NULL,
  conc.colname.col = "Conc.ColName",
  analysis.param.colname = NULL,
  analysis.param.colname.col = "AnalysisParam.ColName",
  additional.colnames = NULL,
  additional.colname.cols = NULL,
  chem.ids,
  chem.lab.id.col = "Chem.Lab.ID",
  chem.name.col = "Compound",
  chem.dtxsid.col = "DTXSID"
)
}
\arguments{
\item{data.label}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{level0.catalog}{A data frame describing which columns of which sheets
in which Excel files contain MS data for analysis. See details for full
explanation.}

\item{date.col}{Which column of input.data indicates the laboratory measurement
date (Defaults to "Date")}

\item{compound.col}{Which column of input.data indicates the test compound
(Defaults to "Compound.Name")}

\item{istd.col}{Which column of input.data indicates the MS peak area for the
internal standard (Defaults to "ISTD.Area")}

\item{dilution.col}{Which column of clint.data indicates how many times the
sample was diluted before MS analysis (Defaults to "Dilution.Factor")}

\item{conc.col}{Which column of clint.data indicates the intended
test chemical concentration 
(units if uM) of
at time zero (Defaults to "Conc")}

\item{time.col}{Which column of clint.data indicates the intended
time of the measurement (in minutes) since the test chemical was introduced
into the hepatocyte incubation (Defaults to "Time")}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}

\item{input.data}{A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{sample.col}{Which column of input.data indicates the unique mass 
spectrometry (MS) sample name used by the laboratory. (Defaults to 
"Lab.Sample.Name")}

\item{density.col}{A single value to be used for all samples indicating
the density of hepatocytes in suspension (10^6 hepatocytes / mL) 
(Defaults to NULL)}

\item{lab.compound.col}{Which column of input.data indicates The test compound 
name used by the laboratory (Defaults to "Lab.Compound.Name")}

\item{dtxsid.col}{Which column of input.data indicates EPA's DSSTox Structure 
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{series.col}{Which column of PPB.data indicates the "series", that is
a simultaneous replicate with the same analytical chemistry 
(Defaults to "Series")}

\item{series}{If this argument is used (defaults to NULL) every observation 
in the table is assigned the value of the argument and the corresponding
column in input.table (if present) is ignored.}

\item{area.col}{Which column of input.data indicates the target analyte (that 
is, the test compound) MS peak area (Defaults to "Area")}

\item{type.col}{Which column of input.data indicates the direction of the 
measurements (either "AtoB" for apical to basolateral or "BtoA" for vice 
versa) (Defaults to "Direction")}

\item{cal.col}{Which column of input.data indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Cal")}

\item{cal}{If this argument is used (defaults to NULL) every observation in
the table is assigned the value of the argument and the corresponding
column in input.table (if present) is ignored.

#param compound.conc.col Which column indicates the intended concentration 
of the test chemical for calibration curves (Defaults to "Standard.Conc")}

\item{dilution}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.name.col}{Which column of input.data indicates identity of the 
internal standard (Defaults to "ISTD.Name")}

\item{istd.name}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{istd.conc.col}{Which column of input.data indicates the concentration of
the internal standard (Defaults to "ISTD.Conc")}

\item{istd.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{nominal.test.conc.col}{Which column of input.data indicates the intended
test chemical concentration at time zero in the dosing solution (added to the
donor side of the Caco-2 test well) (Defaults to "Test.Target.Conc")}

\item{nominal.test.conc}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.method.col}{Which column of input.data indicates the analytical
chemistry analysis method, typically "LCMS" or "GCMS" (Defaults to 
"Analysis.Method")}

\item{analysis.method}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.instrument.col}{Which column of input.data indicates the 
instrument used for chemical analysis, for example 
"Agilent 6890 GC with model 5973 MS" (Defaults to 
"Analysis.Instrument")}

\item{analysis.instrument}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{analysis.parameters.col}{Which column of input.data indicates the 
parameters used to identify the compound on the chemical analysis instrument,
for example 
"Negative Mode, 221.6/161.6, -DPb=26, FPc=-200, EPd=-10, CEe=-20, CXPf=-25.0"
(Defaults to "Analysis.Parameters").}

\item{analysis.parameters}{If this argument is used (defaults to NULL) every 
observation in the table is assigned the value of the argument and the 
corresponding column in input.table (if present) is ignored.}

\item{level0.file.col}{Which column of PPB.data indicates the file from
which the data were obtained (for example "MyWorkbook.xlsx").}

\item{level0.file}{If this argument is used (defaults to NULL) every
observation in the table is assigned the value of the argument and the
corresponding column in input.table (if present) is ignored.}

\item{level0.sheet.col}{Which column of PPB.data indicates the specific 
sheet containing the data if the file is an Excel workbook}

\item{level0.sheet}{If this argument is used (defaults to NULL) every
observation in the table is assigned the value of the argument and the
corresponding column in input.table (if present) is ignored.}
}
\value{
\item{data.frame}{A data.frame in standardized "level1" format}
}
\description{
This function reads multiple Excel files containing mass-spectrometry (MS) data
and extracts the chemical sample data from the specified
sheets. The argument level0.catalog is a table that
provides the necessary information to find the data for each chemical. The
primary data of interest are the analyte peak area, the internal standard
peak area, and the target concentration for calibration curve (CC) samples.
The argument data.label is used to annotate this particular mapping of level0
files into data ready to be organized into a level1 file.
}
\details{
Unless specified to be a single value for all the files, for example sheet="Data",
the argument level0.catalog should be a data frame with the following columns:
\tabular{rr}{
  File \tab The Excel filename to be loaded\cr
  Sheet \tab The name of the Sheet to examine within in the Excel file\cr
  Skip.Rows \tab How many rows should be skipped on the sheet to get usable column names\cr
  Date \tab The date the measurements were made\cr
  Chemical.ID \tab The laboratory chemical identity\cr
  ISTD \tab The internal standard used\cr
  Sample.ColName \tab The column name on the sheet that contains sample identity\cr
  Type.ColName \tab The column name on the sheet that contains the type of sample\cr
  Peak.ColName \tab The column name on the sheet that contains the analyte MS peak area \cr
  ISTD.Peak.ColName \tab The column name on the sheet that contains the internal standard MS peak area\cr
  Conc.ColName \tab The column name on the sheet that contains the intended concentration for calibration curves \cr
  AnalysisParam.ColName \tab The column name on the sheet that contains the MS instrument parameters for the analyte\cr
}
Columns with names ending in ".ColName" indicate the columns to be extracted
from the specified file and sheet.
}
\author{
John Wambaugh
}
