% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fup_uc.R
\name{plot_fup_uc}
\alias{plot_fup_uc}
\title{Plot Mass Spec. Response for Measurement of Fraction Unbound in Plasma (UC)}
\usage{
plot_fup_uc(level2, dtxsid, good.col = "Verified")
}
\arguments{
\item{level2}{A data.frame containing level2 data for fraction unbound in
plasma measured by ultracentrifugation.}

\item{dtxsid}{Which chemical to be plotted.}
}
\value{
\item{ggplot2}{A figure of mass spec. response for different sample types}
}
\description{
This function use describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using ultracentrifugation \insertCite{redgrave1975separation}{invitroTKstats}.
Data are read from a "Level2" text file that should have been formatted and created
by \code{\link{format_fup_red}} (this is the "Level1" file). The Level1 file
should have been curated and had a column added with the value "Y" indicating
that each row is verified as usable for analysis (that is, the Level2 file).
}
\details{
The should be annotated according to
of these types:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
}
}
\author{
John Wambaugh
}
