% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runjagsdata.to.list}
\alias{runjagsdata.to.list}
\title{Convert a runjags object to a list}
\usage{
runjagsdata.to.list(runjagsdata.in)
}
\arguments{
\item{runjagsdata.in}{result from autorun.jags(), an object of class runjags}
}
\value{
a list contains data from the input runjags object
}
\description{
Convert a runjags object to a list
}
