% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_verification.R
\name{sample_verification}
\alias{sample_verification}
\title{Add Sample Verification Column (Level-2)}
\usage{
sample_verification(
  FILENAME,
  data.in,
  exclusion.info,
  assay,
  output.res = TRUE,
  INPUT.DIR = NULL,
  OUTPUT.DIR = NULL
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the output Level-1 file.
"<FILENAME>-<assay>-Level1.tsv".}

\item{data.in}{(Data Frame) A Level-1 data frame from the format functions.}

\item{exclusion.info}{(Data Frame) A data frame containing the variables and 
values of the corresponding variables to exclude rows. 
See details for full explanation.}

\item{assay}{(Character) A string indicating what assay data the input file is. Valid 
input is one of the following: "Clint", "fup-UC", "fup-RED", or "Caco-2". 
This argument only needs to be specified when importing input data set with \code{FILENAME} 
or exporting a data file.}

\item{output.res}{(Logical) When set to \code{TRUE}, the result 
data frame (Level-2) will be exported as a .tsv file to the current directory. 
(Defaults to \code{TRUE}.)}

\item{INPUT.DIR}{(Character) Path to the directory where the input level-1 file exists. 
If \code{NULL}, looking for the input level-1 file in the current working
directory. (Defaults to \code{NULL}.)}

\item{OUTPUT.DIR}{(Character) Path to the directory to save the output file. 
If \code{NULL}, the output file will be saved to the current working
directory or \code{INPUT.DIR} if specified. (Defaults to \code{NULL}.)}
}
\value{
A Level-2 data frame with a verification column.
}
\description{
This function takes in a Level-1 data frame and an exclusion list and 
returns a Level-2 data frame with a verification column. The
verification column contains either "Y", indicating the row is good for analysis,
or messages contained in the exclusion list for why the data rows are excluded. 
If an exclusion list is not provided, all rows are assumed to be good for use 
in further analyses and are verified with "Y".
}
\details{
The `exclusion.info` should be a data frame with the following columns:
\tabular{rr}{
  Variable \tab Level-1 variable used to filter rows for exclusion\cr
  Value \tab Value to exclude\cr
  Message \tab Simple explanation for the exclusion\cr
}
}
\examples{
level2 <- invitroTKstats::kreutz2023.clint

# Data sets in the package are level-2 data. For demonstration purposes,
# remove the verification columns from the data and use it as level-1.
level1 <- dplyr::select(level2, -c(Verifed, Verified))

# scenario 1: input data is in the R session
# create a exclusion criteria data frame
exclusion_criteria <- data.frame(
Variable = c("DTXSID", "Lab.Sample.Name", "Compound.Name"),
Value = c("DTXSID00380798","G4-Ametryn Hep052521 T0a", "Nonafluoropentanamide"), 
Message = c("bad id #1","bad sample", "bad compound")
)

my.level2 <- sample_verification(FILENAME="kreutz", 
data.in=level1, exclusion_criteria, assay="Clint")

# scenario 2: import 'tsv' as input data

# Write the level-1 file to some folder
write.table(data.out,
file="~/invitrotkstats/invitroTKstats/data-raw/kreutz-Clint-Level1.tsv",
sep="\t",
row.names=F,
quote=F)

# import level-1 with INPUT.DIR 
my.level2 <- sample_verification(FILENAME="kreutz", 
exclusion_criteria, assay="Clint", INPUT.DIR = "~/invitrotkstats/invitroTKstats/data-raw")


}
