% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_verification.R
\name{sample_verification}
\alias{sample_verification}
\title{Add Sample Verification Column (Level-2)}
\usage{
sample_verification(
  FILENAME,
  data.in,
  exclusion.info,
  assay,
  output.res = FALSE,
  INPUT.DIR = NULL,
  OUTPUT.DIR = NULL
)
}
\arguments{
\item{FILENAME}{(Character) A string used to identify the output level-1 file.
"<FILENAME>-<assay>-Level1.tsv".}

\item{data.in}{(Data Frame) A level-1 data frame from the format functions.}

\item{exclusion.info}{(Data Frame) A data frame containing the variables and 
values of the corresponding variables to exclude rows. 
See details for full explanation.}

\item{assay}{(Character) A string indicating what assay data the input file is. Valid 
input is one of the following: "Clint", "fup-UC", "fup-RED", or "Caco-2". 
This argument only needs to be specified when importing input data set with \code{FILENAME} 
or exporting a data file.}

\item{output.res}{(Logical) When set to \code{TRUE}, the result 
data frame (level-2) will be exported to the user's per-session temporary directory
or \code{OUTPUT.DIR} (if specified) as a .tsv file. 
(Defaults to \code{FALSE}.)}

\item{INPUT.DIR}{(Character) Path to the directory where the input level-1 file exists. 
If \code{NULL}, looking for the input level-1 file in the current working
directory. (Defaults to \code{NULL}.)}

\item{OUTPUT.DIR}{(Character) Path to the directory to save the output file. 
If \code{NULL}, the output file will be saved to the user's per-session temporary
directory or \code{INPUT.DIR} if specified. (Defaults to \code{NULL}.)}
}
\value{
A level-2 data frame with a verification column.
}
\description{
This function takes in a level-1 data frame and an exclusion list and 
returns a level-2 data frame with a verification column. The
verification column contains either "Y", indicating the row is good for analysis,
or messages contained in the exclusion list for why the data rows are excluded. 
If an exclusion list is not provided, all rows are assumed to be good for use 
in further analyses and are verified with "Y".
}
\details{
The `exclusion.info` should be a data frame with the following columns:
\tabular{rr}{
  Variables \tab level-1 variable(s) used to filter rows for exclusion\cr
  Values \tab Value(s) to exclude\cr
  Message \tab Simple explanation for the exclusion\cr
}
When filtering on multiple variable-value pairs, the character input for 
"Variables" and "Values" should be separated by a vertical bar "|" ,
and the variable-value pairs should match. See demonstration in Examples, Scenario 1. 

If the output level-2 data frame is chosen to be exported and an output directory 
is not specified, it will be exported to the user's R session temporary directory. 
This temporary directory is a per-session directory whose path can be found
with the following code: \code{tempdir()}. For more details, see 
\url{https://www.collinberke.com/til/posts/2023-10-24-temp-directories/}.

As a best practice, \code{INPUT.DIR} (when importing a .tsv file) and/or 
\code{OUTPUT.DIR} should be specified to simplify the process of importing 
and exporting files. This practice ensures that the exported files can easily 
be found and will not be exported to a temporary directory.
}
\examples{
level2 <- invitroTKstats::kreutz2023.clint

# Data sets in the package are level-2 data. For demonstration purposes,
# remove the verification columns from the data and use it as level-1.
level1 <- subset(level2, select = -c(Verifed,Verified) )

# Scenario 1: Pass in data.in and exclusion.info data frame from R session 

# Create a exclusion criteria data frame
# If more than one variable is used to define a set of samples to be excluded,
# enter them as one string, separate the variables with a vertical bar, "|", and do the same for 
# values. 
exclusion_criteria <- data.frame(
  Variables = c(
    "DTXSID","Date","Compound.Name|Lab.Sample.Name"
  ),
  Values = c(
    "DTXSID00380798","21021",
    "Nonafluoropentanamide|Amide Hep120220 T0C2_020921"
  ),
  Message = c(
    "Exclude all samples of this compound.",
    "Exclude all samples from this date.",
    "These samples for this compound were contaminated."
    )
)

# Run the verification function.
my.level2 <- sample_verification(data.in=level1,
                                 exclusion.info = exclusion_criteria,
                                 output.res = FALSE)

# Scenario 2: Import 'tsv' as input data and do not pass in an exclusion.info data frame

\dontrun{
# Write the level-1 file to some folder
# Will need to replace <desired level-1 FOLDER> with desired output folder location  
write.table(level1,
file=here::here("<desired level-1 FOLDER>/kreutz-Clint-Level1.tsv"),
sep="\t",
row.names=F,
quote=F)

# Run the verification function.
# Specify the path to import level-1 data with INPUT.DIR.
# Will need to replace <desired level-1 FOLDER> with chosen output folder location from above 
# If no exclusion.info data frame is used, will label all samples as verified.
# A level-2 file is also saved to INPUT.DIR. 
my.level2 <- sample_verification(FILENAME="kreutz", 
assay="Clint", INPUT.DIR = here::here("<desired level-1 FOLDER>"))
}

}
\author{
Zhihui (Grace) Zhao
}
