% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scientific_10}
\alias{scientific_10}
\title{Formatting function for X-axis in log10-scale}
\usage{
scientific_10(x)
}
\arguments{
\item{x}{}
}
\value{
text with desired expression
}
\description{
Formatting function for X-axis in log10-scale
}
