% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_table.R
\name{summarize_table}
\alias{summarize_table}
\title{Creates a Summary Table of MS Data Frame}
\usage{
summarize_table(
  input.table,
  dtxsid.col = "DTXSID",
  compound.col = "Compound.Name",
  cal.col = "Calibration",
  type.col = "Sample.Type",
  req.types = NULL
)
}
\arguments{
\item{input.table}{(Data Frame) A data frame (Level-1 or Level-2) containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{dtxsid.col}{(Character) Column name of \code{input.table} indicates EPA's DSSTox Structure
ID (\url{http://comptox.epa.gov/dashboard}). (Defaults to "DTXSID".)}

\item{compound.col}{(Character) Column name of \code{input.table} containing the test compound.
(Defaults to "Compound.Name".)}

\item{cal.col}{(Character) Column name of \code{input.table} containing the MS calibration. 
Calibration typically uses indices or dates to represent if the analyses were done on 
different machines on the same day or on different days with the same MS analyzer. 
(Defaults to "Calibration".)}

\item{type.col}{(Character) Column name of \code{input.table} containing the sample type (see tables
in Details). (Defaults to "Sample.Type".)}

\item{req.types}{(Character Vector) A vector of character strings contains
measurement types. If used, the function checks if all of the measurement types included in this vector are
available for each chemical-calibration pair. (Defaults to \code{NULL}.)}
}
\value{
A list containing the summary counts from the input data table. The list includes 
the number of observations, the number of unique chemicals, the number of unique measurements, 
the number of chemicals with complete data sets, the number of chemicals with incomplete data sets, 
and the number of chemicals with repeated observations.
}
\description{
This function creates and returns a list containing summary counts of the input data table.
The list includes the counts of observations, unique chemicals and unique 
measurements in the input data table, and a vector of chemicals that have repeated observations. 
If a vector of data types is specified in the argument \code{req.types}, the function also checks if each chemical has 
observations for every measurement type included in the vector for each chemical-calibration pair. 
If it does, the chemical is said to have a complete data set. Otherwise, it has an incomplete data set. 
The counts of chemicals with complete and incomplete data sets are returned in the output list. 
The input data frame can be Caco-2 data, ultracentrifugation (UC) data, rapid equilibrium dialysis (RED) data, 
or hepatocyte clearance (Clint) data. Tables of measurement types and annotations 
used in each assay are available in Details.
}
\details{
Sample types used in ultracentrifugation (UC) data should be 
annotated according to this table:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultracentrifugation Aqueous Fraction \tab AF\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}

Sample types in rapid equilibrium dialysis (RED) data should be annotated according to
this table:
\tabular{rrrrr}{
  Blank (ignored) \tab Blank\cr
  Plasma well concentration \tab Plasma\cr
  Phosphate-buffered well concentration\tab PBS\cr
  Time zero plasma concentration \tab T0\cr
  Plasma stability sample \tab Stability\cr
  Equilibrium Control Well 1 \tab EC1\cr
  Equilibrium Control Well 2 \tab EC2\cr
  Calibration Curve \tab CC\cr
}

Sample types in hepatocyte clearance data should be annotated according to
this table:
\tabular{rrrrr}{
  Blank \tab Blank\cr
  Hepatocyte incubation concentration \tab Cvst\cr
  Inactivated Hepatocytes \tab Inactive\cr
  Calibration Curve \tab CC\cr
}

Sample types used in Caco-2 data should be annotated according to
this table:
\tabular{rr}{
  Blank with no chemical added \tab Blank \cr
  Dosing vehicle (C0) at target concentration \tab D0\cr
  Donor compartment at end of experiment \tab D2\cr
  Receiver compartment at end of experiment\tab R2\cr
}
}
\examples{
# need to re-visit the path to load data for this example 
load("~/invitroTKstats/Data/Kreutz2023.RData")
summarize_table(kreutz2023.uc, req.types = c("CC", "T1", "T5", "AF"))
summarize_table(kreutz2023.clint)

}
\author{
John Wambaugh
}
