% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_table.R
\name{summarize_table}
\alias{summarize_table}
\title{Creates a standardized data table reporting UC PPB data}
\usage{
summarize_table(
  input.table,
  dtxsid.col = "DTXSID",
  compound.col = "Compound.Name",
  cal.col = "Calibration",
  type.col = "Sample.Type",
  req.types = NULL
)
}
\arguments{
\item{input.table}{A data frame containing mass-spectrometry peak areas,
indication of chemical identity, and measurement type. The data frame should
contain columns with names specified by the following arguments:}

\item{dtxsid.col}{Which column of input.table indicates EPA's DSSTox Structure
ID (\url{http://comptox.epa.gov/dashboard}) (Defaults to "DTXSID")}

\item{compound.col}{Which column of input.table indicates the test compound
(Defaults to "Compound.Name")}

\item{cal.col}{Which column of input.table indicates the MS calibration -- for
instance different machines on the same day or different days with the same
MS analyzer (Defaults to "Calibration")}

\item{type.col}{Which column of input.table indicates the sample type (see table
above)(Defaults to "Sample.Type")}

\item{req.types}{If set (defaults to NULL) all of the measurement types 
included in this vector of character strings must
be available for each chemical-calibration pair to make an estimate.}

\item{FILENAME}{A string used to identify outputs of the function call.
(defaults to "MYDATA")}
}
\value{
\item{list}{A list containing the summary counts from the table}
}
\description{
This function formats data describing mass spectrometry (MS) peak areas
from samples collected as part of in vitro measurement of chemical fraction
unbound in plasma using ultracentrifugation \insertCite{redgrave1975separation}{invitroTKstats}.
An input data frame is organized into a standard set of columns and is written
to a tab-separated text file.
}
\details{
The data frame of observations should be annotated according to
of these types:
\tabular{rrrrr}{
  Calibration Curve \tab CC\cr
  Ultracentrifugation Aqueous Fraction \tab UC\cr
  Whole Plasma T1h Sample  \tab T1\cr
  Whole Plasma T5h Sample \tab T5\cr
}
Chemical concentration is calculated qualitatively as a response:

Response <- AREA / ISTD.AREA * ISTD.CONC
}
\author{
John Wambaugh
}
