% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.clint}
\alias{wambaugh2019.clint}
\title{Mass Spectrometry for \insertCite{wambaugh2019assessing;textual}{invitroTKstats} Hepatocyte Incubations}
\format{
A data.frame 23,021 rows and 15 variables: \describe{
\item{\code{Preferred.Name}}{}
\item{\code{CAS}}{}
\item{\code{DTXSID}}{EPA's DSSTox Structure ID.}
\item{\code{Sample.Name}}{Sample name used by the laboratory.}
\item{\code{Name}}{}
\item{\code{Transition}}{}
\item{\code{Area}}{The test compound MS peak area.}
\item{\code{ISTD.Area}}{MS peak area for the internal standard.}
\item{\code{ISTDResponseRatio}}{}
\item{\code{ln...Remaining}}{}
\item{\code{Time..mins.}}{Time in minues between when the incubation started and when the sample was taken.}
\item{\code{Conc}}{Test compound concentration.}
\item{\code{TaskOrder}}{}
\item{\code{FileName}}{}
\item{\code{X}}{}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.clint
}
\description{
\insertCite{wambaugh2019assessing;textual}{invitroTKstats} includes measurements for intrinsic hepatic clearance
Clint measured using in vitro suspensions of pooled primary human hepatocytes
\insertCite{shibata2002prediction}{invitroTKstats}.
}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}

\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\keyword{data}
