% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.clint}
\alias{wambaugh2019.clint}
\title{Mass Spectrometry for \insertCite{wambaugh2019assessing;textual}{invitroTKstats} Hepatocyte Incubations}
\format{
A data.frame 23,021 rows and 15 variables: \describe{
\item{\code{Preferred.Name}}{}
\item{\code{CAS}}{}
\item{\code{DTXSID}}{}
\item{\code{Sample.Name}}{}
\item{\code{Name}}{}
\item{\code{Transition}}{}
\item{\code{Area}}{}
\item{\code{ISTD.Area}}{}
\item{\code{ISTDResponseRatio}}{}
\item{\code{ln...Remaining}}{}
\item{\code{Time..mins.}}{}
\item{\code{Conc}}{}
\item{\code{TaskOrder}}{}
\item{\code{FileName}}{}
\item{\code{X}}{}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.clint
}
\description{
\insertCite{wambaugh2019assessing;textual}{invitroTKstats} includes measurements for intrinsic hepatic clearance
Clint measured using in vitro suspensions of pooled primary human hepatocytes
\insertCite{shibata2002prediction}{invitroTKstats}.
}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}

\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\keyword{data}
