% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.clint}
\alias{wambaugh2019.clint}
\title{Mass Spectrometry for \insertCite{wambaugh2019assessing;textual}{invitroTKstats} Hepatocyte Incubations}
\format{
A data.frame with 23,021 rows and 11 variables: \describe{
\item{\code{Preferred.Name}}{Preferred compound name from the CompTox Chemicals Dashboard (CCD)}
\item{\code{CAS}}{CAS Registry Number of the test compound}
\item{\code{DTXSID}}{EPA's DSSTox Structure ID}
\item{\code{Sample.Name}}{Sample name used by the laboratory}
\item{\code{Name}}{Compound names used by the laboratory}
\item{\code{Area}}{Mass spectrometry peak area for the test compound}
\item{\code{ISTD.Area}}{Mass spectrometry peak area for the internal standard}
\item{\code{ISTDResponseRatio}}{The ratio between the \code{Area} (area for the tested compound) and the \code{ISTD.Area} (area for the internal standard)}
\item{\code{Time..mins.}}{Time in minutes between the start of incubation and when the sample was taken}
\item{\code{Conc}}{Concentration used in the assay}
\item{\code{FileName}}{Name of the level-0 file containing the data}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.clint
}
\description{
\insertCite{wambaugh2019assessing;textual}{invitroTKstats} includes measurements for intrinsic hepatic clearance
(Clint) measured using in vitro suspensions of pooled primary human hepatocytes
\insertCite{shibata2002prediction}{invitroTKstats}.
}
\references{
\insertRef{shibata2002prediction}{invitroTKstats}

\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\keyword{data}
