% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.red}
\alias{wambaugh2019.red}
\alias{Wambaugh2019.red}
\title{Mass Spectrometry for Wambaugh et al. (2019) Protein Binding}
\format{
A data.frame 17,689 rows and 14 variables: \describe{
\item{\code{Preferred.Name}}{}
\item{\code{CAS}}{}
\item{\code{DTXSID}}{}             
\item{\code{SampleName}}{}
\item{\code{CompoundName}}{}
\item{\code{Transition}}{}          
\item{\code{Area}}{}
\item{\code{ISTD.Area}}{}
\item{\code{ISTDResponseRatio}}{}  
\item{\code{Dilution.Factor}}{}
\item{\code{Protein}}{}
\item{\code{Task.Order}}{}         
\item{\code{RawDataSet}}{}
\item{\code{CyprotexEPASetNumber}}{}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.red
}
\description{
Wambaugh et al. (2019) includes chemical-specific measurments of plasma
protein binding using the method of rapid equilibriuim dialysis (RED, 
Waters et al. 2008)
}
\references{
\insertRef{waters2008validation}{invitroTKstats}
\insertRef{wambaugh2019assessing}{invitroTKstats}
}
\keyword{data}
